Network.AWS.SQS

Description

Amazon Simple Queue Service (SQS) is a fast, reliable, scalable, fully managed message queuing service. SQS makes it simple and cost-effective to decouple the components of a cloud application. You can use SQS to transmit any volume of data, at any level of throughput, without losing messages or requiring other services to be always available. With SQS, you can offload the administrative burden of operating and scaling a highly available messaging cluster, while paying a low price for only what you use.

Documentation

module Network.AWS.SQS.AddPermission

module Network.AWS.SQS.ChangeMessageVisibility

module Network.AWS.SQS.ChangeMessageVisibilityBatch

module Network.AWS.SQS.CreateQueue

module Network.AWS.SQS.DeleteMessage

module Network.AWS.SQS.DeleteMessageBatch

module Network.AWS.SQS.DeleteQueue

module Network.AWS.SQS.GetQueueAttributes

module Network.AWS.SQS.GetQueueUrl

module Network.AWS.SQS.ListDeadLetterSourceQueues

module Network.AWS.SQS.ListQueues

module Network.AWS.SQS.PurgeQueue

module Network.AWS.SQS.ReceiveMessage

module Network.AWS.SQS.RemovePermission

module Network.AWS.SQS.SendMessage

module Network.AWS.SQS.SendMessageBatch

module Network.AWS.SQS.SetQueueAttributes

module Network.AWS.SQS.Types