Network.AWS.SQS.AddPermission

Description

Adds a permission to a queue for a specific principal. This allows for sharing access to the queue. When you create a queue, you have full control access rights for the queue. Only you (as owner of the queue) can grant or deny permissions to the queue. For more information about these permissions, see Shared Queues in the Amazon SQS Developer Guide. &amp;Attribute.1=this &amp;Attribute.2=that.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_AddPermission.html

Request

data AddPermission

Request constructor

addPermission

Arguments

:: Text
apQueueUrl

-> Text
apLabel

-> AddPermission

AddPermission constructor.

The fields accessible through corresponding lenses are:

apAWSAccountIds :: [Text]

apActions :: [Text]

apLabel :: Text

apQueueUrl :: Text

Request lenses

apAWSAccountIds :: Lens' AddPermission [Text]

The AWS account number of the principal who will be given permission. The principal must have an AWS account, but does not need to be signed up for Amazon SQS. For information about locating the AWS account identification, see Your AWS Identifiers in the Amazon SQS Developer Guide.

apActions :: Lens' AddPermission [Text]

The action the client wants to allow for the specified principal. The following are valid values: * | SendMessage | ReceiveMessage | DeleteMessage | ChangeMessageVisibility | GetQueueAttributes | GetQueueUrl. For more information about these actions, see Understanding Permissions in the Amazon SQS Developer Guide. Specifying SendMessage, DeleteMessage, or ChangeMessageVisibility for the ActionName.n also grants permissions for the corresponding batch versions of those actions: SendMessageBatch, DeleteMessageBatch, and ChangeMessageVisibilityBatch.

apLabel :: Lens' AddPermission Text

The unique identification of the permission you're setting (e.g., AliceSendMessage). Constraints: Maximum 80 characters; alphanumeric characters, hyphens (-), and underscores (_) are allowed.

apQueueUrl :: Lens' AddPermission Text

The URL of the Amazon SQS queue to take action on.

Response

data AddPermissionResponse

Response constructor

addPermissionResponse :: AddPermissionResponse

AddPermissionResponse constructor.