Network.AWS.SQS.CreateQueue

Description

Creates a new queue, or returns the URL of an existing one. When you request CreateQueue, you provide a name for the queue. To successfully create a new queue, you must provide a name that is unique within the scope of your own queues.

If you delete a queue, you must wait at least 60 seconds before creating a queue with the same name.

You may pass one or more attributes in the request. If you do not provide a value for any attribute, the queue will have the default value for that attribute. Permitted attributes are the same that can be set using SetQueueAttributes.

Use GetQueueUrl to get a queue's URL. GetQueueUrl requires only the QueueName parameter.

If you provide the name of an existing queue, along with the exact names and values of all the queue's attributes, CreateQueue returns the queue URL for the existing queue. If the queue name, attribute names, or attribute values do not match an existing queue, CreateQueue returns an error.

Some API actions take lists of parameters. These lists are specified using the 'param.n' notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this: '&Attribute.1=this'

'&Attribute.2=that'

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_CreateQueue.html

Request

data CreateQueue

Request constructor

createQueue

Arguments

:: Text
cqQueueName

-> CreateQueue

CreateQueue constructor.

The fields accessible through corresponding lenses are:

cqAttributes :: HashMap Text Text

cqQueueName :: Text

Request lenses

cqAttributes :: Lens' CreateQueue (HashMap Text Text)

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters the CreateQueue action uses:

DelaySeconds - The time in seconds that the delivery of all messages in the queue will be delayed. An integer from 0 to 900 (15 minutes). The default for this attribute is 0 (zero). MaximumMessageSize - The limit of how many bytes a message can contain before Amazon SQS rejects it. An integer from 1024 bytes (1 KiB) up to 262144 bytes (256 KiB). The default for this attribute is 262144 (256 KiB). MessageRetentionPeriod - The number of seconds Amazon SQS retains a message. Integer representing seconds, from 60 (1 minute) to 1209600 (14 days). The default for this attribute is 345600 (4 days). Policy - The queue's policy. A valid AWS policy. For more information about policy structure, see Overview of AWS IAM Policies in the Amazon IAMUser Guide. ReceiveMessageWaitTimeSeconds - The time for which a ReceiveMessage call will wait for a message to arrive. An integer from 0 to 20 (seconds). The default for this attribute is 0. VisibilityTimeout - The visibility timeout for the queue. An integer from 0 to 43200 (12 hours). The default for this attribute is 30. For more information about visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide.

cqQueueName :: Lens' CreateQueue Text

The name for the queue to be created.

Response

data CreateQueueResponse

Response constructor

createQueueResponse :: CreateQueueResponse

CreateQueueResponse constructor.

The fields accessible through corresponding lenses are:

cqrQueueUrl :: Maybe Text

Response lenses

cqrQueueUrl :: Lens' CreateQueueResponse (Maybe Text)

The URL for the created Amazon SQS queue.