Network.AWS.SQS.DeleteMessage

Description

Deletes the specified message from the specified queue. You specify the message by using the message's 'receipt handle' and not the 'message ID' you received when you sent the message. Even if the message is locked by another reader due to the visibility timeout setting, it is still deleted from the queue. If you leave a message in the queue for longer than the queue's configured retention period, Amazon SQS automatically deletes it.

The receipt handle is associated with a specific instance of receiving the message. If you receive a message more than once, the receipt handle you get each time you receive the message is different. When you request DeleteMessage , if you don't provide the most recently received receipt handle for the message, the request will still succeed, but the message might not be deleted.

It is possible you will receive a message even after you have deleted it. This might happen on rare occasions if one of the servers storing a copy of the message is unavailable when you request to delete the message. The copy remains on the server and might be returned to you again on a subsequent receive request. You should create your system to be idempotent so that receiving a particular message more than once is not a problem.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_DeleteMessage.html

Request

data DeleteMessage

Request constructor

deleteMessage

Arguments

:: Text
dmQueueUrl

-> Text
dmReceiptHandle

-> DeleteMessage

DeleteMessage constructor.

The fields accessible through corresponding lenses are:

dmQueueUrl :: Text

dmReceiptHandle :: Text

Request lenses

dmQueueUrl :: Lens' DeleteMessage Text

The URL of the Amazon SQS queue to take action on.

dmReceiptHandle :: Lens' DeleteMessage Text

The receipt handle associated with the message to delete.

Response

data DeleteMessageResponse

Response constructor

deleteMessageResponse :: DeleteMessageResponse

DeleteMessageResponse constructor.