Network.AWS.SQS.GetQueueAttributes

Description

Gets attributes for the specified queue. The following attributes are supported: All - returns all values. ApproximateNumberOfMessages - returns the approximate number of visible messages in a queue. For more information, see Resources Required to Process Messages in the Amazon SQS Developer Guide. ApproximateNumberOfMessagesNotVisible - returns the approximate number of messages that are not timed-out and not deleted. For more information, see Resources Required to Process Messages in the Amazon SQS Developer Guide. VisibilityTimeout - returns the visibility timeout for the queue. For more information about visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide. CreatedTimestamp - returns the time when the queue was created (epoch time in seconds). LastModifiedTimestamp - returns the time when the queue was last changed (epoch time in seconds). Policy - returns the queue's policy. MaximumMessageSize - returns the limit of how many bytes a message can contain before Amazon SQS rejects it. MessageRetentionPeriod - returns the number of seconds Amazon SQS retains a message. QueueArn - returns the queue's Amazon resource name (ARN). ApproximateNumberOfMessagesDelayed - returns the approximate number of messages that are pending to be added to the queue. DelaySeconds - returns the default delay on the queue in seconds. ReceiveMessageWaitTimeSeconds - returns the time for which a ReceiveMessage call will wait for a message to arrive. RedrivePolicy - returns the parameters for dead letter queue functionality of the source queue. For more information about RedrivePolicy and dead letter queues, see Using Amazon SQSDead Letter Queues in the Amazon SQS Developer Guide.

Going forward, new attributes might be added. If you are writing code that calls this action, we recommend that you structure your code so that it can handle new attributes gracefully. Some API actions take lists of parameters. These lists are specified using the 'param.n' notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this: '&Attribute.1=this'

'&Attribute.2=that'

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_GetQueueAttributes.html

Request

data GetQueueAttributes

Request constructor

getQueueAttributes

Arguments

:: Text
gqaQueueUrl

-> GetQueueAttributes

GetQueueAttributes constructor.

The fields accessible through corresponding lenses are:

gqaAttributeNames :: [Text]

gqaQueueUrl :: Text

Request lenses

gqaAttributeNames :: Lens' GetQueueAttributes [Text]

A list of attributes to retrieve information for.

gqaQueueUrl :: Lens' GetQueueAttributes Text

The URL of the Amazon SQS queue to take action on.

Response

data GetQueueAttributesResponse

Response constructor

getQueueAttributesResponse :: GetQueueAttributesResponse

GetQueueAttributesResponse constructor.

The fields accessible through corresponding lenses are:

gqarAttributes :: HashMap Text Text

Response lenses

gqarAttributes :: Lens' GetQueueAttributesResponse (HashMap Text Text)

A map of attributes to the respective values.