Network.AWS.SQS.GetQueueUrl

Description

Returns the URL of an existing queue. This action provides a simple way to retrieve the URL of an Amazon SQS queue.

To access a queue that belongs to another AWS account, use the QueueOwnerAWSAccountId parameter to specify the account ID of the queue's owner. The queue's owner must grant you permission to access the queue. For more information about shared queue access, see AddPermission or go to Shared Queues in the AmazonSQS Developer Guide.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_GetQueueUrl.html

Request

data GetQueueUrl

Request constructor

getQueueUrl

Arguments

:: Text
gquQueueName

-> GetQueueUrl

GetQueueUrl constructor.

The fields accessible through corresponding lenses are:

gquQueueName :: Text

gquQueueOwnerAWSAccountId :: Maybe Text

Request lenses

gquQueueName :: Lens' GetQueueUrl Text

The name of the queue whose URL must be fetched. Maximum 80 characters; alphanumeric characters, hyphens (-), and underscores (_) are allowed.

gquQueueOwnerAWSAccountId :: Lens' GetQueueUrl (Maybe Text)

The AWS account ID of the account that created the queue.

Response

data GetQueueUrlResponse

Response constructor

getQueueUrlResponse

Arguments

:: Text
gqurQueueUrl

-> GetQueueUrlResponse

GetQueueUrlResponse constructor.

The fields accessible through corresponding lenses are:

gqurQueueUrl :: Text

Response lenses

gqurQueueUrl :: Lens' GetQueueUrlResponse Text

The URL for the queue.