Network.AWS.SQS.ListDeadLetterSourceQueues

Description

Returns a list of your queues that have the RedrivePolicy queue attribute configured with a dead letter queue.

For more information about using dead letter queues, see Using Amazon SQSDead Letter Queues.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_ListDeadLetterSourceQueues.html

Request

data ListDeadLetterSourceQueues

Request constructor

listDeadLetterSourceQueues

Arguments

:: Text
ldlsqQueueUrl

-> ListDeadLetterSourceQueues

ListDeadLetterSourceQueues constructor.

The fields accessible through corresponding lenses are:

ldlsqQueueUrl :: Text

Request lenses

ldlsqQueueUrl :: Lens' ListDeadLetterSourceQueues Text

The queue URL of a dead letter queue.

Response

data ListDeadLetterSourceQueuesResponse

Response constructor

listDeadLetterSourceQueuesResponse :: ListDeadLetterSourceQueuesResponse

ListDeadLetterSourceQueuesResponse constructor.

The fields accessible through corresponding lenses are:

ldlsqrQueueUrls :: [Text]

Response lenses

ldlsqrQueueUrls :: Lens' ListDeadLetterSourceQueuesResponse [Text]

A list of source queue URLs that have the RedrivePolicy queue attribute configured with a dead letter queue.