Network.AWS.SQS.ListQueues

Description

Returns a list of your queues. The maximum number of queues that can be returned is 1000. If you specify a value for the optional QueueNamePrefix parameter, only queues with a name beginning with the specified value are returned.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_ListQueues.html

Request

data ListQueues

Request constructor

listQueues :: ListQueues

ListQueues constructor.

The fields accessible through corresponding lenses are:

lqQueueNamePrefix :: Maybe Text

Request lenses

lqQueueNamePrefix :: Lens' ListQueues (Maybe Text)

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

Response

data ListQueuesResponse

Response constructor

listQueuesResponse :: ListQueuesResponse

ListQueuesResponse constructor.

The fields accessible through corresponding lenses are:

lqrQueueUrls :: [Text]

Response lenses

lqrQueueUrls :: Lens' ListQueuesResponse [Text]

A list of queue URLs, up to 1000 entries.