Network.AWS.SQS.PurgeQueue

Description

Deletes the messages in a queue specified by the queue URL.

When you use the PurgeQueue API, the deleted messages in the queue cannot be retrieved. When you purge a queue, the message deletion process takes up to 60 seconds. All messages sent to the queue before calling PurgeQueue will be deleted; messages sent to the queue while it is being purged may be deleted. While the queue is being purged, messages sent to the queue before PurgeQueue was called may be received, but will be deleted within the next minute.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_PurgeQueue.html

Request

data PurgeQueue

Request constructor

purgeQueue

Arguments

:: Text
pqQueueUrl

-> PurgeQueue

PurgeQueue constructor.

The fields accessible through corresponding lenses are:

pqQueueUrl :: Text

Request lenses

pqQueueUrl :: Lens' PurgeQueue Text

The queue URL of the queue to delete the messages from when using the PurgeQueue API.

Response

data PurgeQueueResponse

Response constructor

purgeQueueResponse :: PurgeQueueResponse

PurgeQueueResponse constructor.