Network.AWS.SQS.RemovePermission

Description

Revokes any permissions in the queue policy that matches the specified Label parameter. Only the owner of the queue can remove permissions.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_RemovePermission.html

Request

data RemovePermission

Request constructor

removePermission

Arguments

:: Text
rpQueueUrl

-> Text
rpLabel

-> RemovePermission

RemovePermission constructor.

The fields accessible through corresponding lenses are:

rpLabel :: Text

rpQueueUrl :: Text

Request lenses

rpLabel :: Lens' RemovePermission Text

The identification of the permission to remove. This is the label added with the AddPermission action.

rpQueueUrl :: Lens' RemovePermission Text

The URL of the Amazon SQS queue to take action on.

Response

data RemovePermissionResponse

Response constructor

removePermissionResponse :: RemovePermissionResponse

RemovePermissionResponse constructor.