Network.AWS.SQS.SendMessage

Description

Delivers a message to the specified queue. With Amazon SQS, you now have the ability to send large payload messages that are up to 256KB (262,144 bytes) in size. To send large payloads, you must use an AWS SDK that supports SigV4 signing. To verify whether SigV4 is supported for an AWS SDK, check the SDK release notes.

The following list shows the characters (in Unicode) allowed in your message, according to the W3C XML specification. For more information, go to http://www.w3.org/TR/REC-xml/#charsets If you send any characters not included in the list, your request will be rejected.

xA | x20 to xE000 to x10000 to #x10FFFF]

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_SendMessage.html

Request

data SendMessage

Request constructor

sendMessage

Arguments

:: Text
smQueueUrl

-> Text
smMessageBody

-> SendMessage

SendMessage constructor.

The fields accessible through corresponding lenses are:

smDelaySeconds :: Maybe Int

smMessageAttributes :: HashMap Text MessageAttributeValue

smMessageBody :: Text

smQueueUrl :: Text

Request lenses

smDelaySeconds :: Lens' SendMessage (Maybe Int)

The number of seconds (0 to 900 - 15 minutes) to delay a specific message. Messages with a positive DelaySeconds value become available for processing after the delay time is finished. If you don't specify a value, the default value for the queue applies.

smMessageAttributes :: Lens' SendMessage (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

smMessageBody :: Lens' SendMessage Text

The message to send. String maximum 256 KB in size. For a list of allowed characters, see the preceding important note.

smQueueUrl :: Lens' SendMessage Text

The URL of the Amazon SQS queue to take action on.

Response

data SendMessageResponse

Response constructor

sendMessageResponse :: SendMessageResponse

SendMessageResponse constructor.

The fields accessible through corresponding lenses are:

smrMD5OfMessageAttributes :: Maybe Text

smrMD5OfMessageBody :: Maybe Text

smrMessageId :: Maybe Text

Response lenses

smrMD5OfMessageAttributes :: Lens' SendMessageResponse (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http://www.faqs.org/rfcs/rfc1321.html.

smrMD5OfMessageBody :: Lens' SendMessageResponse (Maybe Text)

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http://www.faqs.org/rfcs/rfc1321.html.

smrMessageId :: Lens' SendMessageResponse (Maybe Text)

An element containing the message ID of the message sent to the queue. For more information, see Queue and Message Identifiers in the Amazon SQSDeveloper Guide.