Network.AWS.SQS.SendMessageBatch

Description

Delivers up to ten messages to the specified queue. This is a batch version of SendMessage. The result of the send action on each message is reported individually in the response. The maximum allowed individual message size is 256 KB (262,144 bytes).

The maximum total payload size (i.e., the sum of all a batch's individual message lengths) is also 256 KB (262,144 bytes).

If the DelaySeconds parameter is not specified for an entry, the default for the queue is used.

The following list shows the characters (in Unicode) that are allowed in your message, according to the W3C XML specification. For more information, go to http://www.faqs.org/rfcs/rfc1321.html. If you send any characters that are not included in the list, your request will be rejected. xA | #xD | [xD7FF] | [xFFFD] | [x10FFFF]

Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of 200. Some API actions take lists of parameters. These lists are specified using the 'param.n' notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this: '&Attribute.1=this'

'&Attribute.2=that'

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_SendMessageBatch.html

Request

data SendMessageBatch

Request constructor

sendMessageBatch

Arguments

:: Text
smbQueueUrl

-> SendMessageBatch

SendMessageBatch constructor.

The fields accessible through corresponding lenses are:

smbEntries :: [SendMessageBatchRequestEntry]

smbQueueUrl :: Text

Request lenses

smbEntries :: Lens' SendMessageBatch [SendMessageBatchRequestEntry]

A list of SendMessageBatchRequestEntry items.

smbQueueUrl :: Lens' SendMessageBatch Text

The URL of the Amazon SQS queue to take action on.

Response

data SendMessageBatchResponse

Response constructor

sendMessageBatchResponse :: SendMessageBatchResponse

SendMessageBatchResponse constructor.

The fields accessible through corresponding lenses are:

smbrFailed :: [BatchResultErrorEntry]

smbrSuccessful :: [SendMessageBatchResultEntry]

Response lenses

smbrFailed :: Lens' SendMessageBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items with the error detail about each message that could not be enqueued.

smbrSuccessful :: Lens' SendMessageBatchResponse [SendMessageBatchResultEntry]

A list of SendMessageBatchResultEntry items.