Network.AWS.SQS.SetQueueAttributes

Description

Sets the value of one or more queue attributes. When you change a queue's attributes, the change can take up to 60 seconds for most of the attributes to propagate throughout the SQS system. Changes made to the MessageRetentionPeriod attribute can take up to 15 minutes.

Going forward, new attributes might be added. If you are writing code that calls this action, we recommend that you structure your code so that it can handle new attributes gracefully.

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_SetQueueAttributes.html

Request

data SetQueueAttributes

Request constructor

setQueueAttributes

Arguments

:: Text
sqaQueueUrl

-> SetQueueAttributes

SetQueueAttributes constructor.

The fields accessible through corresponding lenses are:

sqaAttributes :: HashMap Text Text

sqaQueueUrl :: Text

Request lenses

sqaAttributes :: Lens' SetQueueAttributes (HashMap Text Text)

A map of attributes to set.

The following lists the names, descriptions, and values of the special request parameters the SetQueueAttributes action uses:

DelaySeconds - The time in seconds that the delivery of all messages in the queue will be delayed. An integer from 0 to 900 (15 minutes). The default for this attribute is 0 (zero). MaximumMessageSize - The limit of how many bytes a message can contain before Amazon SQS rejects it. An integer from 1024 bytes (1 KiB) up to 262144 bytes (256 KiB). The default for this attribute is 262144 (256 KiB). MessageRetentionPeriod - The number of seconds Amazon SQS retains a message. Integer representing seconds, from 60 (1 minute) to 1209600 (14 days). The default for this attribute is 345600 (4 days). Policy - The queue's policy. A valid AWS policy. For more information about policy structure, see Overview of AWS IAM Policies in the Amazon IAMUser Guide. ReceiveMessageWaitTimeSeconds - The time for which a ReceiveMessage call will wait for a message to arrive. An integer from 0 to 20 (seconds). The default for this attribute is 0. VisibilityTimeout - The visibility timeout for the queue. An integer from 0 to 43200 (12 hours). The default for this attribute is 30. For more information about visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide. RedrivePolicy - The parameters for dead letter queue functionality of the source queue. For more information about RedrivePolicy and dead letter queues, see Using Amazon SQS Dead Letter Queues in the Amazon SQS Developer Guide.

sqaQueueUrl :: Lens' SetQueueAttributes Text

The URL of the Amazon SQS queue to take action on.

Response

data SetQueueAttributesResponse

Response constructor

setQueueAttributesResponse :: SetQueueAttributesResponse

SetQueueAttributesResponse constructor.