Network.AWS.SQS.Types

Description

Amazon Simple Queue Service (SQS) is a fast, reliable, scalable, fully managed message queuing service. SQS makes it simple and cost-effective to decouple the components of a cloud application. You can use SQS to transmit any volume of data, at any level of throughput, without losing messages or requiring other services to be always available. With SQS, you can offload the administrative burden of operating and scaling a highly available messaging cluster, while paying a low price for only what you use.

Service

data SQS

Supported version (2012-11-05) of the Amazon Simple Queue Service service.

Errors

data SQSError

A sum type representing possible errors returned by the SQS service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

BatchEntryIdsNotDistinct
Two or more batch entries have the same Id in the request.

BatchRequestTooLong
The length of all the messages put together is more than the limit.

EmptyBatchRequest
Batch request does not contain an entry.

InvalidAttributeName
The attribute referred to does not exist.

InvalidBatchEntryId
The Id of a batch entry in a batch request does not abide by the specification.

InvalidIdFormat
The receipt handle is not valid for the current version.

InvalidMessageContents
The message contains characters outside the allowed set.

MessageNotInflight
The message referred to is not in flight.

OverLimit
The action that you requested would violate a limit. For example, ReceiveMessage returns this error if the maximum number of messages inflight has already been reached. AddPermission returns this error if the maximum number of permissions for the queue has already been reached.

QueueDeletedRecently
You must wait 60 seconds after deleting a queue before you can create another with the same name.

QueueDoesNotExist
The queue referred to does not exist.

QueueNameExists
A queue already exists with this name. Amazon SQS returns this error only if the request includes attributes whose values differ from those of the existing queue.

ReceiptHandleIsInvalid
The receipt handle provided is not valid.

SQSClient HttpException

SQSSerializer String

SQSService String

TooManyEntriesInBatchRequest
Batch request contains more number of entries than permissible.

UnsupportedOperation
Error code 400. Unsupported operation.

_BatchEntryIdsNotDistinct :: Prism' SQSError ()

Two or more batch entries have the same Id in the request.

See: BatchEntryIdsNotDistinct

_BatchRequestTooLong :: Prism' SQSError ()

The length of all the messages put together is more than the limit.

See: BatchRequestTooLong

_EmptyBatchRequest :: Prism' SQSError ()

Batch request does not contain an entry.

See: EmptyBatchRequest

_InvalidAttributeName :: Prism' SQSError ()

The attribute referred to does not exist.

See: InvalidAttributeName

_InvalidBatchEntryId :: Prism' SQSError ()

The Id of a batch entry in a batch request does not abide by the specification.

See: InvalidBatchEntryId

_InvalidIdFormat :: Prism' SQSError ()

The receipt handle is not valid for the current version.

See: InvalidIdFormat

_InvalidMessageContents :: Prism' SQSError ()

The message contains characters outside the allowed set.

See: InvalidMessageContents

_MessageNotInflight :: Prism' SQSError ()

The message referred to is not in flight.

See: MessageNotInflight

_OverLimit :: Prism' SQSError ()

The action that you requested would violate a limit. For example, ReceiveMessage returns this error if the maximum number of messages inflight has already been reached. AddPermission returns this error if the maximum number of permissions for the queue has already been reached.

See: OverLimit

_QueueDeletedRecently :: Prism' SQSError ()

You must wait 60 seconds after deleting a queue before you can create another with the same name.

See: QueueDeletedRecently

_QueueDoesNotExist :: Prism' SQSError ()

The queue referred to does not exist.

See: QueueDoesNotExist

_QueueNameExists :: Prism' SQSError ()

A queue already exists with this name. Amazon SQS returns this error only if the request includes attributes whose values differ from those of the existing queue.

See: QueueNameExists

_ReceiptHandleIsInvalid :: Prism' SQSError ()

The receipt handle provided is not valid.

See: ReceiptHandleIsInvalid

_SQSClient :: Prism' SQSError HttpException

See: SQSClient

_SQSSerializer :: Prism' SQSError String

See: SQSSerializer

_SQSService :: Prism' SQSError String

See: SQSService

_TooManyEntriesInBatchRequest :: Prism' SQSError ()

Batch request contains more number of entries than permissible.

See: TooManyEntriesInBatchRequest

_UnsupportedOperation :: Prism' SQSError ()

Error code 400. Unsupported operation.

See: UnsupportedOperation

XML

xmlOptions :: Tagged a XMLOptions

QueueAttributeName

data QueueAttributeName

Constructors

QueueAttributeNameApproximateNumberOfMessages
ApproximateNumberOfMessages

QueueAttributeNameApproximateNumberOfMessagesDelayed
ApproximateNumberOfMessagesDelayed

QueueAttributeNameApproximateNumberOfMessagesNotVisible
ApproximateNumberOfMessagesNotVisible

QueueAttributeNameCreatedTimestamp
CreatedTimestamp

QueueAttributeNameDelaySeconds
DelaySeconds

QueueAttributeNameLastModifiedTimestamp
LastModifiedTimestamp

QueueAttributeNameMaximumMessageSize
MaximumMessageSize

QueueAttributeNameMessageRetentionPeriod
MessageRetentionPeriod

QueueAttributeNamePolicy
Policy

QueueAttributeNameQueueArn
QueueArn

QueueAttributeNameReceiveMessageWaitTimeSeconds
ReceiveMessageWaitTimeSeconds

QueueAttributeNameRedrivePolicy
RedrivePolicy

QueueAttributeNameVisibilityTimeout
VisibilityTimeout

ChangeMessageVisibilityBatchResultEntry

data ChangeMessageVisibilityBatchResultEntry

Encloses the id of an entry in ChangeMessageVisibilityBatch.

changeMessageVisibilityBatchResultEntry

Arguments

:: Text
cmvbrerId

-> ChangeMessageVisibilityBatchResultEntry

Smart constructor for the minimum required fields to construct a valid ChangeMessageVisibilityBatchResultEntry data type.

ChangeMessageVisibilityBatchResultEntry is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Text

cmvbrerId :: Lens' ChangeMessageVisibilityBatchResultEntry Text

Represents a message whose visibility timeout has been changed successfully.

DeleteMessageBatchResultEntry

data DeleteMessageBatchResultEntry

Encloses the id an entry in DeleteMessageBatch.

deleteMessageBatchResultEntry

Arguments

:: Text
dmbrerId

-> DeleteMessageBatchResultEntry

Smart constructor for the minimum required fields to construct a valid DeleteMessageBatchResultEntry data type.

DeleteMessageBatchResultEntry is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Text

dmbrerId :: Lens' DeleteMessageBatchResultEntry Text

Represents a successfully deleted message.

BatchResultErrorEntry

data BatchResultErrorEntry

This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

batchResultErrorEntry

Arguments

:: Text
breeId

-> Bool
breeSenderFault

-> Text
breeCode

-> BatchResultErrorEntry

Smart constructor for the minimum required fields to construct a valid BatchResultErrorEntry data type.

BatchResultErrorEntry is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Text

SenderFault :: Bool

Code :: Text

Message :: Maybe Text

breeId :: Lens' BatchResultErrorEntry Text

The id of an entry in a batch request.

breeSenderFault :: Lens' BatchResultErrorEntry Bool

Whether the error happened due to the sender's fault.

breeCode :: Lens' BatchResultErrorEntry Text

An error code representing why the action failed on this entry.

breeMessage :: Lens' BatchResultErrorEntry (Maybe Text)

A message explaining why the action failed on this entry.

ChangeMessageVisibilityBatchRequestEntry

data ChangeMessageVisibilityBatchRequestEntry

Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch. All of the following parameters are list parameters that must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this: &amp;ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2 &amp;ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=Your_Receipt_Handle &amp;ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45.

changeMessageVisibilityBatchRequestEntry

Arguments

:: Text
cmvbreId

-> Text
cmvbreReceiptHandle

-> ChangeMessageVisibilityBatchRequestEntry

Smart constructor for the minimum required fields to construct a valid ChangeMessageVisibilityBatchRequestEntry data type to populate a request.

The fields accessible through corresponding lenses are:

Id :: Text

ReceiptHandle :: Text

VisibilityTimeout :: Maybe Integer

cmvbreId :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

cmvbreReceiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

A receipt handle.

cmvbreVisibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Integer)

The new value (in seconds) for the message's visibility timeout.

DeleteMessageBatchRequestEntry

data DeleteMessageBatchRequestEntry

Encloses a receipt handle and an identifier for it.

deleteMessageBatchRequestEntry

Arguments

:: Text
dmbreId

-> Text
dmbreReceiptHandle

-> DeleteMessageBatchRequestEntry

Smart constructor for the minimum required fields to construct a valid DeleteMessageBatchRequestEntry data type to populate a request.

The fields accessible through corresponding lenses are:

Id :: Text

ReceiptHandle :: Text

dmbreId :: Lens' DeleteMessageBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

dmbreReceiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

A receipt handle.

Message

data Message

An Amazon SQS message.

message :: Message

Smart constructor for the minimum required fields to construct a valid Message data type.

Message is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

MessageId :: Maybe Text

ReceiptHandle :: Maybe Text

MD5OfBody :: Maybe Text

Body :: Maybe Text

Attributes :: Map QueueAttributeName Text

MD5OfMessageAttributes :: Maybe Text

MessageAttributes :: Map Text MessageAttributeValue

mMessageId :: Lens' Message (Maybe Text)

A unique identifier for the message. Message IDs are considered unique across all AWS accounts for an extended period of time.

mReceiptHandle :: Lens' Message (Maybe Text)

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

mMD5OfBody :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message body string.

mBody :: Lens' Message (Maybe Text)

The message's contents (not URL-encoded).

mAttributes :: Lens' Message (Map QueueAttributeName Text)

SenderId, SentTimestamp, ApproximateReceiveCount, and/or ApproximateFirstReceiveTimestamp. SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the epoch time in milliseconds.

mMD5OfMessageAttributes :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

mMessageAttributes :: Lens' Message (Map Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see SendMessage. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes).

messageAttributeValue

Arguments

:: Text
mavDataType

-> MessageAttributeValue

Smart constructor for the minimum required fields to construct a valid MessageAttributeValue data type to populate a request.

The fields accessible through corresponding lenses are:

StringValue :: Maybe Text

BinaryValue :: Maybe ByteString

StringListValues :: [Text]

BinaryListValues :: [ByteString]

DataType :: Text

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters.

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

mavStringListValues :: Lens' MessageAttributeValue [Text]

Not implemented. Reserved for future use.

mavBinaryListValues :: Lens' MessageAttributeValue [ByteString]

Not implemented. Reserved for future use.

mavDataType :: Lens' MessageAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. In addition, you can append your own custom labels. For more information, see Message Attribute Data Types.

SendMessageBatchRequestEntry

data SendMessageBatchRequestEntry

Contains the details of a single Amazon SQS message along with a Id.

sendMessageBatchRequestEntry

Arguments

:: Text
smbreId

-> Text
smbreMessageBody

-> SendMessageBatchRequestEntry

Smart constructor for the minimum required fields to construct a valid SendMessageBatchRequestEntry data type to populate a request.

The fields accessible through corresponding lenses are:

Id :: Text

MessageBody :: Text

DelaySeconds :: Maybe Integer

MessageAttributes :: Map Text MessageAttributeValue

smbreId :: Lens' SendMessageBatchRequestEntry Text

An identifier for the message in this batch. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

smbreMessageBody :: Lens' SendMessageBatchRequestEntry Text

Body of the message.

smbreDelaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Integer)

The number of seconds for which the message has to be delayed.

smbreMessageAttributes :: Lens' SendMessageBatchRequestEntry (Map Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

SendMessageBatchResultEntry

data SendMessageBatchResultEntry

Encloses a message ID for successfully enqueued message of a SendMessageBatch.

sendMessageBatchResultEntry

Arguments

:: Text
smbrerId

-> Text
smbrerMessageId

-> Text
smbrerMD5OfMessageBody

-> SendMessageBatchResultEntry

Smart constructor for the minimum required fields to construct a valid SendMessageBatchResultEntry data type.

SendMessageBatchResultEntry is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Text

MessageId :: Text

MD5OfMessageBody :: Text

MD5OfMessageAttributes :: Maybe Text

smbrerId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message in this batch.

smbrerMessageId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message.

smbrerMD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

smbrerMD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message batch correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.