Network.AWS.SQS.V2012_11_05

Description

Amazon Simple Queue Service (SQS) is a fast, reliable, scalable, fully managed message queuing service. SQS makes it simple and cost-effective to decouple the components of a cloud application. You can use SQS to transmit any volume of data, at any level of throughput, without losing messages or requiring other services to be always available. With SQS, you can offload the administrative burden of operating and scaling a highly available messaging cluster, while paying a low price for only what you use.

Documentation

data AddPermission

Constructors

AddPermission

Fields

_aprAWSAccountIds :: [Text]

The AWS account number of the principal who will be given permission. The principal must have an AWS account, but does not need to be signed up for Amazon SQS. For information about locating the AWS account identification, see Your AWS Identifiers in the Amazon SQS Developer Guide.

_aprActions :: [Text]

The action the client wants to allow for the specified principal. The following are valid values: * | SendMessage | ReceiveMessage | DeleteMessage | ChangeMessageVisibility | GetQueueAttributes | GetQueueUrl. For more information about these actions, see Understanding Permissions in the Amazon SQS Developer Guide. Specifying SendMessage, DeleteMessage, or ChangeMessageVisibility for the ActionName.n also grants permissions for the corresponding batch versions of those actions: SendMessageBatch, DeleteMessageBatch, and ChangeMessageVisibilityBatch.

_aprQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_aprLabel :: Text

The unique identification of the permission you're setting (e.g., AliceSendMessage). Constraints: Maximum 80 characters; alphanumeric characters, hyphens (-), and underscores (_) are allowed.

aprQueueUrl :: Lens' AddPermission Text

aprLabel :: Lens' AddPermission Text

aprActions :: Lens' AddPermission [Text]

aprAWSAccountIds :: Lens' AddPermission [Text]

data AddPermissionResponse

Constructors

AddPermissionResponse

data ChangeMessageVisibility

Constructors

ChangeMessageVisibility

Fields

_cmvrVisibilityTimeout :: Integer

The new value (in seconds - from 0 to 43200 - maximum 12 hours) for the message's visibility timeout.

_cmvrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_cmvrReceiptHandle :: Text

The receipt handle associated with the message whose visibility timeout should be changed. This parameter is returned by the ReceiveMessage action.

cmvrVisibilityTimeout :: Lens' ChangeMessageVisibility Integer

cmvrReceiptHandle :: Lens' ChangeMessageVisibility Text

cmvrQueueUrl :: Lens' ChangeMessageVisibility Text

data ChangeMessageVisibilityResponse

Constructors

ChangeMessageVisibilityResponse

data ChangeMessageVisibilityBatch

Constructors

ChangeMessageVisibilityBatch

Fields

_cmvbrEntries :: [ChangeMessageVisibilityBatchRequestEntry]

A list of receipt handles of the messages for which the visibility timeout must be changed.

_cmvbrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

cmvbrQueueUrl :: Lens' ChangeMessageVisibilityBatch Text

cmvbrEntries :: Lens' ChangeMessageVisibilityBatch [ChangeMessageVisibilityBatchRequestEntry]

data ChangeMessageVisibilityBatchResponse

Constructors

ChangeMessageVisibilityBatchResponse

Fields

_cmvbsFailed :: [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.

_cmvbsSuccessful :: [ChangeMessageVisibilityBatchResultEntry]

A list of ChangeMessageVisibilityBatchResultEntry items.

cmvbsSuccessful :: Lens' ChangeMessageVisibilityBatchResponse [ChangeMessageVisibilityBatchResultEntry]

cmvbsFailed :: Lens' ChangeMessageVisibilityBatchResponse [BatchResultErrorEntry]

createQueue

Arguments

:: Text
_cqrQueueName

-> CreateQueue

Minimum specification for a CreateQueue request.

data CreateQueue

Constructors

CreateQueue

Fields

_cqrQueueName :: Text

The name for the queue to be created.

_cqrAttributes :: Map QueueAttributeName Text

A map of attributes with their corresponding values. The following lists the names, descriptions, and values of the special request parameters the CreateQueue action uses: DelaySeconds - The time in seconds that the delivery of all messages in the queue will be delayed. An integer from 0 to 900 (15 minutes). The default for this attribute is 0 (zero). MaximumMessageSize - The limit of how many bytes a message can contain before Amazon SQS rejects it. An integer from 1024 bytes (1 KiB) up to 262144 bytes (256 KiB). The default for this attribute is 262144 (256 KiB). MessageRetentionPeriod - The number of seconds Amazon SQS retains a message. Integer representing seconds, from 60 (1 minute) to 1209600 (14 days). The default for this attribute is 345600 (4 days). Policy - The queue's policy. A valid form-url-encoded policy. For more information about policy structure, see Basic Policy Structure in the Amazon SQS Developer Guide. For more information about form-url-encoding, see http:/www.w3.orgMarkUphtml-spechtml-spec_8.html#SEC8.2.1. ReceiveMessageWaitTimeSeconds - The time for which a ReceiveMessage call will wait for a message to arrive. An integer from 0 to 20 (seconds). The default for this attribute is 0. VisibilityTimeout - The visibility timeout for the queue. An integer from 0 to 43200 (12 hours). The default for this attribute is 30. For more information about visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide.

cqrQueueName :: Lens' CreateQueue Text

cqrAttributes :: Lens' CreateQueue (Map QueueAttributeName Text)

data CreateQueueResponse

Constructors

CreateQueueResponse

Fields

_cqsQueueUrl :: Maybe Text

The URL for the created Amazon SQS queue.

cqsQueueUrl :: Iso' CreateQueueResponse (Maybe Text)

data DeleteMessage

Constructors

DeleteMessage

Fields

_dmrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_dmrReceiptHandle :: Text

The receipt handle associated with the message to delete.

dmrReceiptHandle :: Lens' DeleteMessage Text

dmrQueueUrl :: Lens' DeleteMessage Text

data DeleteMessageResponse

Constructors

DeleteMessageResponse

data DeleteMessageBatch

Constructors

DeleteMessageBatch

Fields

_dmbrEntries :: [DeleteMessageBatchRequestEntry]

A list of receipt handles for the messages to be deleted.

_dmbrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

dmbrQueueUrl :: Lens' DeleteMessageBatch Text

dmbrEntries :: Lens' DeleteMessageBatch [DeleteMessageBatchRequestEntry]

data DeleteMessageBatchResponse

Constructors

DeleteMessageBatchResponse

Fields

_dmbsFailed :: [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.

_dmbsSuccessful :: [DeleteMessageBatchResultEntry]

A list of DeleteMessageBatchResultEntry items.

dmbsSuccessful :: Lens' DeleteMessageBatchResponse [DeleteMessageBatchResultEntry]

dmbsFailed :: Lens' DeleteMessageBatchResponse [BatchResultErrorEntry]

data DeleteQueue

Constructors

DeleteQueue

Fields

_dqrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

dqrQueueUrl :: Iso' DeleteQueue Text

data DeleteQueueResponse

Constructors

DeleteQueueResponse

getQueueAttributes

Arguments

:: Text
_gqarQueueUrl

-> GetQueueAttributes

Minimum specification for a GetQueueAttributes request.

data GetQueueAttributes

Constructors

GetQueueAttributes

Fields

_gqarQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_gqarAttributeNames :: [QueueAttributeName]

A list of attributes to retrieve information for.

gqarQueueUrl :: Lens' GetQueueAttributes Text

gqarAttributeNames :: Lens' GetQueueAttributes [QueueAttributeName]

data GetQueueAttributesResponse

Constructors

GetQueueAttributesResponse

Fields

_gqasAttributes :: Map QueueAttributeName Text

A map of attributes to the respective values.

gqasAttributes :: Iso' GetQueueAttributesResponse (Map QueueAttributeName Text)

getQueueUrl

Arguments

:: Text
_gqurQueueName

-> GetQueueUrl

Minimum specification for a GetQueueUrl request.

data GetQueueUrl

Constructors

GetQueueUrl

Fields

_gqurQueueName :: Text

The name of the queue whose URL must be fetched. Maximum 80 characters; alphanumeric characters, hyphens (-), and underscores (_) are allowed.

_gqurQueueOwnerAWSAccountId :: Maybe Text

The AWS account ID of the account that created the queue.

gqurQueueOwnerAWSAccountId :: Lens' GetQueueUrl (Maybe Text)

gqurQueueName :: Lens' GetQueueUrl Text

data GetQueueUrlResponse

Constructors

GetQueueUrlResponse

Fields

_gqusQueueUrl :: Maybe Text

The URL for the queue.

gqusQueueUrl :: Iso' GetQueueUrlResponse (Maybe Text)

data ListDeadLetterSourceQueues

Constructors

ListDeadLetterSourceQueues

Fields

_ldlsqrQueueUrl :: Text

The queue URL of a dead letter queue.

ldlsqrQueueUrl :: Iso' ListDeadLetterSourceQueues Text

data ListDeadLetterSourceQueuesResponse

Constructors

ListDeadLetterSourceQueuesResponse

Fields

_ldlsqsQueueUrls :: [Text]

A list of source queue URLs that have the RedrivePolicy queue attribute configured with a dead letter queue.

ldlsqsQueueUrls :: Iso' ListDeadLetterSourceQueuesResponse [Text]

listQueues :: ListQueues

Minimum specification for a ListQueues request.

data ListQueues

Constructors

ListQueues

Fields

_lqrQueueNamePrefix :: Maybe Text

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

lqrQueueNamePrefix :: Iso' ListQueues (Maybe Text)

data ListQueuesResponse

Constructors

ListQueuesResponse

Fields

_lqsQueueUrls :: [Text]

A list of queue URLs, up to 1000 entries.

lqsQueueUrls :: Iso' ListQueuesResponse [Text]

receiveMessage

Arguments

:: Text
_rmrQueueUrl

-> ReceiveMessage

Minimum specification for a ReceiveMessage request.

data ReceiveMessage

Constructors

ReceiveMessage

Fields

_rmrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_rmrAttributeNames :: [QueueAttributeName]

A list of attributes that need to be returned along with each message. The following lists the names and descriptions of the attributes that can be returned: All - returns all values. ApproximateFirstReceiveTimestamp - returns the time when the message was first received (epoch time in milliseconds). ApproximateReceiveCount - returns the number of times a message has been received but not deleted. SenderId - returns the AWS account number (or the IP address, if anonymous access is allowed) of the sender. SentTimestamp - returns the time when the message was sent (epoch time in milliseconds).

_rmrVisibilityTimeout :: Maybe Integer

The duration (in seconds) that the received messages are hidden from subsequent retrieve requests after being retrieved by a ReceiveMessage request.

_rmrWaitTimeSeconds :: Maybe Integer

The duration (in seconds) for which the call will wait for a message to arrive in the queue before returning. If a message is available, the call will return sooner than WaitTimeSeconds.

_rmrMaxNumberOfMessages :: Maybe Integer

The maximum number of messages to return. Amazon SQS never returns more messages than this value but may return fewer. Values can be from 1 to 10. Default is 1. All of the messages are not necessarily returned.

_rmrMessageAttributeNames :: [Text]

The message attribute Name can contain the following characters: A-Z, a-z, 0-9, underscore(_), hyphen(-), and period (.). The message attribute name must not start or end with a period, and it should not have successive periods. The message attribute name is case sensitive and must be unique among all attribute names for the message. The message attribute name can be up to 256 characters long. Attribute names cannot start with "AWS." or "Amazon." because these prefixes are reserved for use by Amazon Web Services.

rmrWaitTimeSeconds :: Lens' ReceiveMessage (Maybe Integer)

rmrVisibilityTimeout :: Lens' ReceiveMessage (Maybe Integer)

rmrQueueUrl :: Lens' ReceiveMessage Text

rmrMessageAttributeNames :: Lens' ReceiveMessage [Text]

rmrMaxNumberOfMessages :: Lens' ReceiveMessage (Maybe Integer)

rmrAttributeNames :: Lens' ReceiveMessage [QueueAttributeName]

data ReceiveMessageResponse

Constructors

ReceiveMessageResponse

Fields

_rmsMessages :: [Message]

A list of messages.

rmsMessages :: Iso' ReceiveMessageResponse [Message]

data RemovePermission

Constructors

RemovePermission

Fields

_rprQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_rprLabel :: Text

The identification of the permission to remove. This is the label added with the AddPermission action.

rprQueueUrl :: Lens' RemovePermission Text

rprLabel :: Lens' RemovePermission Text

data RemovePermissionResponse

Constructors

RemovePermissionResponse

sendMessage

Arguments

:: Text
_smrQueueUrl

-> Text
_smrMessageBody

-> SendMessage

Minimum specification for a SendMessage request.

data SendMessage

Constructors

SendMessage

Fields

_smrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_smrMessageBody :: Text

The message to send. String maximum 256 KB in size. For a list of allowed characters, see the preceding important note.

_smrDelaySeconds :: Maybe Integer

The number of seconds (0 to 900 - 15 minutes) to delay a specific message. Messages with a positive DelaySeconds value become available for processing after the delay time is finished. If you don't specify a value, the default value for the queue applies.

_smrMessageAttributes :: Map Text MessageAttributeValue

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

smrQueueUrl :: Lens' SendMessage Text

smrMessageBody :: Lens' SendMessage Text

smrMessageAttributes :: Lens' SendMessage (Map Text MessageAttributeValue)

smrDelaySeconds :: Lens' SendMessage (Maybe Integer)

data SendMessageResponse

Constructors

SendMessageResponse

Fields

_smsMessageId :: Maybe Text

An element containing the message ID of the message sent to the queue. For more information, see Queue and Message Identifiers in the Amazon SQS Developer Guide.

_smsMD5OfMessageBody :: Maybe Text

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

_smsMD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

smsMessageId :: Lens' SendMessageResponse (Maybe Text)

smsMD5OfMessageBody :: Lens' SendMessageResponse (Maybe Text)

smsMD5OfMessageAttributes :: Lens' SendMessageResponse (Maybe Text)

data SendMessageBatch

Constructors

SendMessageBatch

Fields

_smbrEntries :: [SendMessageBatchRequestEntry]

A list of SendMessageBatchRequestEntry items.

_smbrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

smbrQueueUrl :: Lens' SendMessageBatch Text

smbrEntries :: Lens' SendMessageBatch [SendMessageBatchRequestEntry]

data SendMessageBatchResponse

Constructors

SendMessageBatchResponse

Fields

_smbsFailed :: [BatchResultErrorEntry]

A list of BatchResultErrorEntry items with the error detail about each message that could not be enqueued.

_smbsSuccessful :: [SendMessageBatchResultEntry]

A list of SendMessageBatchResultEntry items.

smbsSuccessful :: Lens' SendMessageBatchResponse [SendMessageBatchResultEntry]

smbsFailed :: Lens' SendMessageBatchResponse [BatchResultErrorEntry]

data SetQueueAttributes

Constructors

SetQueueAttributes

Fields

_sqarAttributes :: Map QueueAttributeName Text

A map of attributes to set. The following lists the names, descriptions, and values of the special request parameters the SetQueueAttributes action uses: DelaySeconds - The time in seconds that the delivery of all messages in the queue will be delayed. An integer from 0 to 900 (15 minutes). The default for this attribute is 0 (zero). MaximumMessageSize - The limit of how many bytes a message can contain before Amazon SQS rejects it. An integer from 1024 bytes (1 KiB) up to 262144 bytes (256 KiB). The default for this attribute is 262144 (256 KiB). MessageRetentionPeriod - The number of seconds Amazon SQS retains a message. Integer representing seconds, from 60 (1 minute) to 1209600 (14 days). The default for this attribute is 345600 (4 days). Policy - The queue's policy. A valid form-url-encoded policy. For more information about policy structure, see Basic Policy Structure in the Amazon SQS Developer Guide. For more information about form-url-encoding, see http:/www.w3.orgMarkUphtml-spechtml-spec_8.html#SEC8.2.1. ReceiveMessageWaitTimeSeconds - The time for which a ReceiveMessage call will wait for a message to arrive. An integer from 0 to 20 (seconds). The default for this attribute is 0. VisibilityTimeout - The visibility timeout for the queue. An integer from 0 to 43200 (12 hours). The default for this attribute is 30. For more information about visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide. RedrivePolicy - The parameters for dead letter queue functionality of the source queue. For more information about RedrivePolicy and dead letter queues, see Using Amazon SQS Dead Letter Queues in the Amazon SQS Developer Guide.

_sqarQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

sqarQueueUrl :: Lens' SetQueueAttributes Text

sqarAttributes :: Lens' SetQueueAttributes (Map QueueAttributeName Text)

data SetQueueAttributesResponse

Constructors

SetQueueAttributesResponse

data SQS

Supported version (2012-11-05) of the Amazon Simple Queue Service service.

xmlOptions :: Tagged a XMLOptions

data QueueAttributeName

Constructors

QueueAttributeNameApproximateNumberOfMessages
ApproximateNumberOfMessages

QueueAttributeNameApproximateNumberOfMessagesDelayed
ApproximateNumberOfMessagesDelayed

QueueAttributeNameApproximateNumberOfMessagesNotVisible
ApproximateNumberOfMessagesNotVisible

QueueAttributeNameCreatedTimestamp
CreatedTimestamp

QueueAttributeNameDelaySeconds
DelaySeconds

QueueAttributeNameLastModifiedTimestamp
LastModifiedTimestamp

QueueAttributeNameMaximumMessageSize
MaximumMessageSize

QueueAttributeNameMessageRetentionPeriod
MessageRetentionPeriod

QueueAttributeNamePolicy
Policy

QueueAttributeNameQueueArn
QueueArn

QueueAttributeNameReceiveMessageWaitTimeSeconds
ReceiveMessageWaitTimeSeconds

QueueAttributeNameRedrivePolicy
RedrivePolicy

QueueAttributeNameVisibilityTimeout
VisibilityTimeout

newtype ChangeMessageVisibilityBatchResultEntry

Encloses the id of an entry in ChangeMessageVisibilityBatch.

Constructors

ChangeMessageVisibilityBatchResultEntry

Fields

_cmvbrfId :: Text

Represents a message whose visibility timeout has been changed successfully.

newtype DeleteMessageBatchResultEntry

Encloses the id an entry in DeleteMessageBatch.

Constructors

DeleteMessageBatchResultEntry

Fields

_dmbrfId :: Text

Represents a successfully deleted message.

data BatchResultErrorEntry

This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

Constructors

BatchResultErrorEntry

Fields

_breeSenderFault :: Bool

Whether the error happened due to the sender's fault.

_breeId :: Text

The id of an entry in a batch request.

_breeCode :: Text

An error code representing why the action failed on this entry.

_breeMessage :: Maybe Text

A message explaining why the action failed on this entry.

data ChangeMessageVisibilityBatchRequestEntry

Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch. All of the following parameters are list parameters that must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this: &amp;ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2 &amp;ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=Your_Receipt_Handle &amp;ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45.

Constructors

ChangeMessageVisibilityBatchRequestEntry

Fields

_cmvbreVisibilityTimeout :: Maybe Integer

The new value (in seconds) for the message's visibility timeout.

_cmvbreId :: Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

_cmvbreReceiptHandle :: Text

A receipt handle.

data DeleteMessageBatchRequestEntry

Encloses a receipt handle and an identifier for it.

Constructors

DeleteMessageBatchRequestEntry

Fields

_dmbreId :: Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

_dmbreReceiptHandle :: Text

A receipt handle.

data Message

An Amazon SQS message.

Constructors

Message

Fields

_nMessageAttributes :: Map Text MessageAttributeValue

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

_nMD5OfBody :: Maybe Text

An MD5 digest of the non-URL-encoded message body string.

_nBody :: Maybe Text

The message's contents (not URL-encoded).

_nAttributes :: Map QueueAttributeName Text

SenderId, SentTimestamp, ApproximateReceiveCount, and/or ApproximateFirstReceiveTimestamp. SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the epoch time in milliseconds.

_nReceiptHandle :: Maybe Text

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

_nMessageId :: Maybe Text

A unique identifier for the message. Message IDs are considered unique across all AWS accounts for an extended period of time.

_nMD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see SendMessage. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes).

Constructors

MessageAttributeValue

Fields

_mavBinaryValue :: Maybe ByteString

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

_mavStringListValues :: [Text]

Not implemented. Reserved for future use.

_mavStringValue :: Maybe Text

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters.

_mavBinaryListValues :: [ByteString]

Not implemented. Reserved for future use.

_mavDataType :: Text

Amazon SQS supports the following logical data types: String, Number, and Binary. In addition, you can append your own custom labels. For more information, see Message Attribute Data Types.

data SendMessageBatchRequestEntry

Contains the details of a single Amazon SQS message along with a Id.

Constructors

SendMessageBatchRequestEntry

Fields

_smbreMessageAttributes :: Map Text MessageAttributeValue

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

_smbreDelaySeconds :: Maybe Integer

The number of seconds for which the message has to be delayed.

_smbreId :: Text

An identifier for the message in this batch. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

_smbreMessageBody :: Text

Body of the message.

data SendMessageBatchResultEntry

Encloses a message ID for successfully enqueued message of a SendMessageBatch.

Constructors

SendMessageBatchResultEntry

Fields

_smbrfId :: Text

An identifier for the message in this batch.

_smbrfMessageId :: Text

An identifier for the message.

_smbrfMD5OfMessageBody :: Text

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

_smbrfMD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message batch correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

cmvbrfId :: Iso' ChangeMessageVisibilityBatchResultEntry Text

dmbrfId :: Iso' DeleteMessageBatchResultEntry Text

breeSenderFault :: Lens' BatchResultErrorEntry Bool

breeMessage :: Lens' BatchResultErrorEntry (Maybe Text)

breeId :: Lens' BatchResultErrorEntry Text

breeCode :: Lens' BatchResultErrorEntry Text

cmvbreVisibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Integer)

cmvbreReceiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

cmvbreId :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

dmbreReceiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

dmbreId :: Lens' DeleteMessageBatchRequestEntry Text

nReceiptHandle :: Lens' Message (Maybe Text)

nMessageId :: Lens' Message (Maybe Text)

nMessageAttributes :: Lens' Message (Map Text MessageAttributeValue)

nMD5OfMessageAttributes :: Lens' Message (Maybe Text)

nMD5OfBody :: Lens' Message (Maybe Text)

nBody :: Lens' Message (Maybe Text)

nAttributes :: Lens' Message (Map QueueAttributeName Text)

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

mavStringListValues :: Lens' MessageAttributeValue [Text]

mavDataType :: Lens' MessageAttributeValue Text

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

mavBinaryListValues :: Lens' MessageAttributeValue [ByteString]

smbreMessageBody :: Lens' SendMessageBatchRequestEntry Text

smbreMessageAttributes :: Lens' SendMessageBatchRequestEntry (Map Text MessageAttributeValue)

smbreId :: Lens' SendMessageBatchRequestEntry Text

smbreDelaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Integer)

smbrfMessageId :: Lens' SendMessageBatchResultEntry Text

smbrfMD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

smbrfMD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

smbrfId :: Lens' SendMessageBatchResultEntry Text