Network.AWS.SQS.V2012_11_05.GetQueueUrl

Description

Returns the URL of an existing queue. This action provides a simple way to retrieve the URL of an Amazon SQS queue. To access a queue that belongs to another AWS account, use the QueueOwnerAWSAccountId parameter to specify the account ID of the queue's owner. The queue's owner must grant you permission to access the queue. For more information about shared queue access, see AddPermission or go to Shared Queues in the Amazon SQS Developer Guide. The following example Query request gets the URL for the specified queue. http:/sqs.us-east-1.amazonaws.com ?Action=GetQueueUrl &QueueName=testQueue &Version=2011-10-01 &SignatureMethod=HmacSHA256 &Expires=2011-10-24T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE http:/&useast1-query;123456789012/testQueue 470a6f13-2ed9-4181-ad8a-2fdea142988e.

Documentation

getQueueUrl

Arguments

:: Text
_gqurQueueName

-> GetQueueUrl

Minimum specification for a GetQueueUrl request.

data GetQueueUrl

Constructors

GetQueueUrl

Fields

_gqurQueueName :: Text

The name of the queue whose URL must be fetched. Maximum 80 characters; alphanumeric characters, hyphens (-), and underscores (_) are allowed.

_gqurQueueOwnerAWSAccountId :: Maybe Text

The AWS account ID of the account that created the queue.

gqurQueueOwnerAWSAccountId :: Lens' GetQueueUrl (Maybe Text)

gqurQueueName :: Lens' GetQueueUrl Text

data GetQueueUrlResponse

Constructors

GetQueueUrlResponse

Fields

_gqusQueueUrl :: Maybe Text

The URL for the queue.

gqusQueueUrl :: Iso' GetQueueUrlResponse (Maybe Text)