Network.AWS.SQS.V2012_11_05.ListDeadLetterSourceQueues

Description

Returns a list of your queues that have the RedrivePolicy queue attribute configured with a dead letter queue. The following example Query request returns a list of dead letter source queues. In this example only one source queue, MySourceQueue, was configured with a dead letter queue. Action=ListDeadLetterSourceQueues &Version=2012-11-05 http:/sqs.us-east-1.amazonaws.com123456789012/MySourceQueue 8ffb921f-b85e-53d9-abcf-d8d0057f38fc For more information about using dead letter queues, see Using Amazon SQS Dead Letter Queues.

Request

data ListDeadLetterSourceQueues

Request constructor

mkListDeadLetterSourceQueues

Arguments

:: Text
ldlsqQueueUrl

-> ListDeadLetterSourceQueues

Smart constructor for the minimum required parameters to construct a valid ListDeadLetterSourceQueues request.

The fields accessible through corresponding lenses are:

QueueUrl :: Text

Request lenses

ldlsqQueueUrl :: Lens' ListDeadLetterSourceQueues Text

The queue URL of a dead letter queue.

Response

data ListDeadLetterSourceQueuesResponse

A list of your dead letter source queues.

Response constructor

mkListDeadLetterSourceQueuesResponse

Arguments

:: [Text]
ldlsqrQueueUrls

-> ListDeadLetterSourceQueuesResponse

Smart constructor for the minimum required parameters to construct a valid ListDeadLetterSourceQueuesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

QueueUrls :: [Text]

Response lenses

ldlsqrQueueUrls :: Lens' ListDeadLetterSourceQueuesResponse [Text]

A list of source queue URLs that have the RedrivePolicy queue attribute configured with a dead letter queue.