Network.AWS.SQS.V2012_11_05.ListQueues

Description

Returns a list of your queues. The maximum number of queues that can be returned is 1000. If you specify a value for the optional QueueNamePrefix parameter, only queues with a name beginning with the specified value are returned. The following example Query request returns the queues whose names begin with the letter T. http:/sqs.us-east-1.amazonaws.com ?Action=ListQueues &QueueNamePrefix=t &Version=2009-02-01 &SignatureMethod=HmacSHA256 &Expires=2009-04-18T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE http:/sqs.us-east-1.amazonaws.com123456789012/testQueue 725275ae-0b9b-4762-b238-436d7c65a1ac.

Request

data ListQueues

Request constructor

mkListQueues :: ListQueues

Smart constructor for the minimum required parameters to construct a valid ListQueues request.

Request lenses

lqQueueNamePrefix :: Lens' ListQueues (Maybe Text)

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

Response

data ListQueuesResponse

A list of your queues.

Response lenses

lqrsQueueUrls :: Lens' ListQueuesResponse [Text]

A list of queue URLs, up to 1000 entries.