Network.AWS.SQS.V2012_11_05.Types

Description

Amazon Simple Queue Service (SQS) is a fast, reliable, scalable, fully managed message queuing service. SQS makes it simple and cost-effective to decouple the components of a cloud application. You can use SQS to transmit any volume of data, at any level of throughput, without losing messages or requiring other services to be always available. With SQS, you can offload the administrative burden of operating and scaling a highly available messaging cluster, while paying a low price for only what you use.

Service

data SQS

Supported version (2012-11-05) of the Amazon Simple Queue Service service.

Errors

data family Er a

XML

xmlOptions :: Tagged a XMLOptions

QueueAttributeName

data QueueAttributeName

Constructors

QueueAttributeNameApproximateNumberOfMessages
ApproximateNumberOfMessages

QueueAttributeNameApproximateNumberOfMessagesDelayed
ApproximateNumberOfMessagesDelayed

QueueAttributeNameApproximateNumberOfMessagesNotVisible
ApproximateNumberOfMessagesNotVisible

QueueAttributeNameCreatedTimestamp
CreatedTimestamp

QueueAttributeNameDelaySeconds
DelaySeconds

QueueAttributeNameLastModifiedTimestamp
LastModifiedTimestamp

QueueAttributeNameMaximumMessageSize
MaximumMessageSize

QueueAttributeNameMessageRetentionPeriod
MessageRetentionPeriod

QueueAttributeNamePolicy
Policy

QueueAttributeNameQueueArn
QueueArn

QueueAttributeNameReceiveMessageWaitTimeSeconds
ReceiveMessageWaitTimeSeconds

QueueAttributeNameRedrivePolicy
RedrivePolicy

QueueAttributeNameVisibilityTimeout
VisibilityTimeout

ChangeMessageVisibilityBatchResultEntry

data ChangeMessageVisibilityBatchResultEntry

Encloses the id of an entry in ChangeMessageVisibilityBatch.

mkChangeMessageVisibilityBatchResultEntry

Arguments

:: Text
cmvbrersId

-> ChangeMessageVisibilityBatchResultEntry

Smart constructor for the minimum required fields to construct a valid ChangeMessageVisibilityBatchResultEntry data type to populate a request.

cmvbrersId :: Lens' ChangeMessageVisibilityBatchResultEntry Text

Represents a message whose visibility timeout has been changed successfully.

DeleteMessageBatchResultEntry

data DeleteMessageBatchResultEntry

Encloses the id an entry in DeleteMessageBatch.

mkDeleteMessageBatchResultEntry

Arguments

:: Text
dmbrersId

-> DeleteMessageBatchResultEntry

Smart constructor for the minimum required fields to construct a valid DeleteMessageBatchResultEntry data type to populate a request.

dmbrersId :: Lens' DeleteMessageBatchResultEntry Text

Represents a successfully deleted message.

BatchResultErrorEntry

data BatchResultErrorEntry

This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

mkBatchResultErrorEntry

Arguments

:: Text
breeId

-> Bool
breeSenderFault

-> Text
breeCode

-> BatchResultErrorEntry

Smart constructor for the minimum required fields to construct a valid BatchResultErrorEntry data type to populate a request.

breeId :: Lens' BatchResultErrorEntry Text

The id of an entry in a batch request.

breeSenderFault :: Lens' BatchResultErrorEntry Bool

Whether the error happened due to the sender's fault.

breeCode :: Lens' BatchResultErrorEntry Text

An error code representing why the action failed on this entry.

breeMessage :: Lens' BatchResultErrorEntry (Maybe Text)

A message explaining why the action failed on this entry.

ChangeMessageVisibilityBatchRequestEntry

data ChangeMessageVisibilityBatchRequestEntry

Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch. All of the following parameters are list parameters that must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this: &amp;ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2 &amp;ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=Your_Receipt_Handle &amp;ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45.

mkChangeMessageVisibilityBatchRequestEntry

Arguments

:: Text
cmvbreId

-> Text
cmvbreReceiptHandle

-> ChangeMessageVisibilityBatchRequestEntry

Smart constructor for the minimum required fields to construct a valid ChangeMessageVisibilityBatchRequestEntry data type to populate a request.

cmvbreId :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

cmvbreReceiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

A receipt handle.

cmvbreVisibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Integer)

The new value (in seconds) for the message's visibility timeout.

DeleteMessageBatchRequestEntry

data DeleteMessageBatchRequestEntry

Encloses a receipt handle and an identifier for it.

mkDeleteMessageBatchRequestEntry

Arguments

:: Text
dmbreId

-> Text
dmbreReceiptHandle

-> DeleteMessageBatchRequestEntry

Smart constructor for the minimum required fields to construct a valid DeleteMessageBatchRequestEntry data type to populate a request.

dmbreId :: Lens' DeleteMessageBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

dmbreReceiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

A receipt handle.

Message

data Message

An Amazon SQS message.

mkMessage :: Message

Smart constructor for the minimum required fields to construct a valid Message data type to populate a request.

mMessageId :: Lens' Message (Maybe Text)

A unique identifier for the message. Message IDs are considered unique across all AWS accounts for an extended period of time.

mReceiptHandle :: Lens' Message (Maybe Text)

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

mMD5OfBody :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message body string.

mBody :: Lens' Message (Maybe Text)

The message's contents (not URL-encoded).

mAttributes :: Lens' Message (Map QueueAttributeName Text)

SenderId, SentTimestamp, ApproximateReceiveCount, and/or ApproximateFirstReceiveTimestamp. SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the epoch time in milliseconds.

mMD5OfMessageAttributes :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

mMessageAttributes :: Lens' Message (Map Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see SendMessage. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes).

mkMessageAttributeValue

Arguments

:: Text
mavDataType

-> MessageAttributeValue

Smart constructor for the minimum required fields to construct a valid MessageAttributeValue data type to populate a request.

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters.

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

mavStringListValues :: Lens' MessageAttributeValue [Text]

Not implemented. Reserved for future use.

mavBinaryListValues :: Lens' MessageAttributeValue [ByteString]

Not implemented. Reserved for future use.

mavDataType :: Lens' MessageAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. In addition, you can append your own custom labels. For more information, see Message Attribute Data Types.

SendMessageBatchRequestEntry

data SendMessageBatchRequestEntry

Contains the details of a single Amazon SQS message along with a Id.

mkSendMessageBatchRequestEntry

Arguments

:: Text
smbreId

-> Text
smbreMessageBody

-> SendMessageBatchRequestEntry

Smart constructor for the minimum required fields to construct a valid SendMessageBatchRequestEntry data type to populate a request.

smbreId :: Lens' SendMessageBatchRequestEntry Text

An identifier for the message in this batch. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

smbreMessageBody :: Lens' SendMessageBatchRequestEntry Text

Body of the message.

smbreDelaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Integer)

The number of seconds for which the message has to be delayed.

smbreMessageAttributes :: Lens' SendMessageBatchRequestEntry (Map Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

SendMessageBatchResultEntry

data SendMessageBatchResultEntry

Encloses a message ID for successfully enqueued message of a SendMessageBatch.

mkSendMessageBatchResultEntry

Arguments

:: Text
smbrersId

-> Text
smbrersMessageId

-> Text
smbrersMD5OfMessageBody

-> SendMessageBatchResultEntry

Smart constructor for the minimum required fields to construct a valid SendMessageBatchResultEntry data type to populate a request.

smbrersId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message in this batch.

smbrersMessageId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message.

smbrersMD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

smbrersMD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message batch correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.