Network.AWS.SQS.AddPermission

Description

Adds a permission to a queue for a specific principal . This allows sharing access to the queue.

When you create a queue, you have full control access rights for the queue. Only you, the owner of the queue, can grant or deny permissions to the queue. For more information about these permissions, see Shared Queues in the Amazon SQS Developer Guide .

Creating a Request

addPermission

Arguments

:: Text
apQueueURL

-> Text
apLabel

-> AddPermission

Creates a value of AddPermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apQueueURL - The URL of the Amazon SQS queue to which permissions are added. Queue URLs are case-sensitive.

apLabel - The unique identification of the permission you're setting (for example, AliceSendMessage ). Maximum 80 characters. Allowed characters include alphanumeric characters, hyphens (- ), and underscores (_ ).

apAWSAccountIds - The AWS account number of the principal who is given permission. The principal must have an AWS account, but does not need to be signed up for Amazon SQS. For information about locating the AWS account identification, see Your AWS Identifiers in the Amazon SQS Developer Guide .

apActions - The action the client wants to allow for the specified principal. The following values are valid: * * * ChangeMessageVisibility * DeleteMessage * GetQueueAttributes * GetQueueUrl * ReceiveMessage * SendMessage For more information about these actions, see Understanding Permissions in the Amazon SQS Developer Guide . Specifying SendMessage , DeleteMessage , or ChangeMessageVisibility for ActionName.n also grants permissions for the corresponding batch versions of those actions: SendMessageBatch , DeleteMessageBatch , and ChangeMessageVisibilityBatch .

data AddPermission

See: addPermission smart constructor.

Request Lenses

apQueueURL :: Lens' AddPermission Text

The URL of the Amazon SQS queue to which permissions are added. Queue URLs are case-sensitive.

apLabel :: Lens' AddPermission Text

The unique identification of the permission you're setting (for example, AliceSendMessage ). Maximum 80 characters. Allowed characters include alphanumeric characters, hyphens (- ), and underscores (_ ).

apAWSAccountIds :: Lens' AddPermission [Text]

The AWS account number of the principal who is given permission. The principal must have an AWS account, but does not need to be signed up for Amazon SQS. For information about locating the AWS account identification, see Your AWS Identifiers in the Amazon SQS Developer Guide .

apActions :: Lens' AddPermission [Text]

The action the client wants to allow for the specified principal. The following values are valid: * * * ChangeMessageVisibility * DeleteMessage * GetQueueAttributes * GetQueueUrl * ReceiveMessage * SendMessage For more information about these actions, see Understanding Permissions in the Amazon SQS Developer Guide . Specifying SendMessage , DeleteMessage , or ChangeMessageVisibility for ActionName.n also grants permissions for the corresponding batch versions of those actions: SendMessageBatch , DeleteMessageBatch , and ChangeMessageVisibilityBatch .

Destructuring the Response

addPermissionResponse :: AddPermissionResponse

Creates a value of AddPermissionResponse with the minimum fields required to make a request.

data AddPermissionResponse

See: addPermissionResponse smart constructor.