Network.AWS.SQS.AddPermission

Description

Adds a permission to a queue for a specific principal. This allows for sharing access to the queue.

When you create a queue, you have full control access rights for the queue. Only you (as owner of the queue) can grant or deny permissions to the queue. For more information about these permissions, see Shared Queues in the AmazonSQS Developer Guide.

AddPermission writes an Amazon SQS-generated policy. If you want to write your own policy, use SetQueueAttributes to upload your policy. For more information about writing your own policy, see Using The Access PolicyLanguage in the Amazon SQS Developer Guide.

Some API actions take lists of parameters. These lists are specified using the 'param.n' notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this: '&Attribute.1=this'

'&Attribute.2=that'

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_AddPermission.html

Request

data AddPermission

Request constructor

addPermission

Arguments

:: Text
apQueueUrl

-> Text
apLabel

-> AddPermission

AddPermission constructor.

The fields accessible through corresponding lenses are:

apAWSAccountIds :: [Text]

apActions :: [Text]

apLabel :: Text

apQueueUrl :: Text

Request lenses

apAWSAccountIds :: Lens' AddPermission [Text]

The AWS account number of the principal who will be given permission. The principal must have an AWS account, but does not need to be signed up for Amazon SQS. For information about locating the AWS account identification, see Your AWS Identifiers in the Amazon SQS Developer Guide.

apActions :: Lens' AddPermission [Text]

The action the client wants to allow for the specified principal. The following are valid values: '* | SendMessage | ReceiveMessage | DeleteMessage| ChangeMessageVisibility | GetQueueAttributes | GetQueueUrl'. For more information about these actions, see Understanding Permissions in the AmazonSQS Developer Guide.

Specifying SendMessage, DeleteMessage, or ChangeMessageVisibility for the n also grants permissions for the corresponding batch versions of those actions: SendMessageBatch, DeleteMessageBatch, and ChangeMessageVisibilityBatch.

apLabel :: Lens' AddPermission Text

The unique identification of the permission you're setting (e.g., AliceSendMessage). Constraints: Maximum 80 characters; alphanumeric characters, hyphens (-), and underscores (_) are allowed.

apQueueUrl :: Lens' AddPermission Text

The URL of the Amazon SQS queue to take action on.

Response

data AddPermissionResponse

Response constructor

addPermissionResponse :: AddPermissionResponse

AddPermissionResponse constructor.