Network.AWS.SQS.ChangeMessageVisibilityBatch

Description

Changes the visibility timeout of multiple messages. This is a batch version of ChangeMessageVisibility . The result of the action on each message is reported individually in the response. You can send up to 10 ChangeMessageVisibility requests with each ChangeMessageVisibilityBatch action.

Important: Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of 200 .

Creating a Request

changeMessageVisibilityBatch

Arguments

:: Text
cmvbQueueURL

-> ChangeMessageVisibilityBatch

Creates a value of ChangeMessageVisibilityBatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmvbQueueURL - The URL of the Amazon SQS queue whose messages' visibility is changed. Queue URLs are case-sensitive.

cmvbEntries - A list of receipt handles of the messages for which the visibility timeout must be changed.

data ChangeMessageVisibilityBatch

See: changeMessageVisibilityBatch smart constructor.

Request Lenses

cmvbQueueURL :: Lens' ChangeMessageVisibilityBatch Text

The URL of the Amazon SQS queue whose messages' visibility is changed. Queue URLs are case-sensitive.

cmvbEntries :: Lens' ChangeMessageVisibilityBatch [ChangeMessageVisibilityBatchRequestEntry]

A list of receipt handles of the messages for which the visibility timeout must be changed.

Destructuring the Response

changeMessageVisibilityBatchResponse

Arguments

:: Int
cmvbrsResponseStatus

-> ChangeMessageVisibilityBatchResponse

Creates a value of ChangeMessageVisibilityBatchResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmvbrsResponseStatus - -- | The response status code.

cmvbrsSuccessful - A list of ChangeMessageVisibilityBatchResultEntry items.

cmvbrsFailed - A list of BatchResultErrorEntry items.

data ChangeMessageVisibilityBatchResponse

For each message in the batch, the response contains a ChangeMessageVisibilityBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

See: changeMessageVisibilityBatchResponse smart constructor.

Response Lenses

cmvbrsResponseStatus :: Lens' ChangeMessageVisibilityBatchResponse Int

- | The response status code.

cmvbrsSuccessful :: Lens' ChangeMessageVisibilityBatchResponse [ChangeMessageVisibilityBatchResultEntry]

A list of ChangeMessageVisibilityBatchResultEntry items.

cmvbrsFailed :: Lens' ChangeMessageVisibilityBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.