Network.AWS.SQS.ChangeMessageVisibilityBatch

Description

Changes the visibility timeout of multiple messages. This is a batch version of ChangeMessageVisibility. The result of the action on each message is reported individually in the response. You can send up to 10 ChangeMessageVisibility requests with each ChangeMessageVisibilityBatch action.

Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of 200. Some API actions take lists of parameters. These lists are specified using the 'param.n' notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this: '&Attribute.1=this'

'&Attribute.2=that'

http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_ChangeMessageVisibilityBatch.html

Request

data ChangeMessageVisibilityBatch

Request constructor

changeMessageVisibilityBatch

Arguments

:: Text
cmvbQueueUrl

-> ChangeMessageVisibilityBatch

ChangeMessageVisibilityBatch constructor.

The fields accessible through corresponding lenses are:

cmvbEntries :: [ChangeMessageVisibilityBatchRequestEntry]

cmvbQueueUrl :: Text

Request lenses

cmvbEntries :: Lens' ChangeMessageVisibilityBatch [ChangeMessageVisibilityBatchRequestEntry]

A list of receipt handles of the messages for which the visibility timeout must be changed.

cmvbQueueUrl :: Lens' ChangeMessageVisibilityBatch Text

The URL of the Amazon SQS queue to take action on.

Response

data ChangeMessageVisibilityBatchResponse

Response constructor

changeMessageVisibilityBatchResponse :: ChangeMessageVisibilityBatchResponse

ChangeMessageVisibilityBatchResponse constructor.

The fields accessible through corresponding lenses are:

cmvbrFailed :: [BatchResultErrorEntry]

cmvbrSuccessful :: [ChangeMessageVisibilityBatchResultEntry]

Response lenses

cmvbrFailed :: Lens' ChangeMessageVisibilityBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.

cmvbrSuccessful :: Lens' ChangeMessageVisibilityBatchResponse [ChangeMessageVisibilityBatchResultEntry]

A list of ChangeMessageVisibilityBatchResultEntry items.