Network.AWS.SQS.DeleteMessage

Description

Deletes the specified message from the specified queue. You specify the message by using the message's receipt handle and not the MessageId you receive when you send the message. Even if the message is locked by another reader due to the visibility timeout setting, it is still deleted from the queue. If you leave a message in the queue for longer than the queue's configured retention period, Amazon SQS automatically deletes the message.

Creating a Request

deleteMessage

Arguments

:: Text
dmQueueURL

-> Text
dmReceiptHandle

-> DeleteMessage

Creates a value of DeleteMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmQueueURL - The URL of the Amazon SQS queue from which messages are deleted. Queue URLs are case-sensitive.

dmReceiptHandle - The receipt handle associated with the message to delete.

data DeleteMessage

See: deleteMessage smart constructor.

Request Lenses

dmQueueURL :: Lens' DeleteMessage Text

The URL of the Amazon SQS queue from which messages are deleted. Queue URLs are case-sensitive.

dmReceiptHandle :: Lens' DeleteMessage Text

The receipt handle associated with the message to delete.

Destructuring the Response

deleteMessageResponse :: DeleteMessageResponse

Creates a value of DeleteMessageResponse with the minimum fields required to make a request.

data DeleteMessageResponse

See: deleteMessageResponse smart constructor.