Network.AWS.SQS.DeleteMessageBatch

Description

Deletes up to ten messages from the specified queue. This is a batch version of DeleteMessage . The result of the action on each message is reported individually in the response.

Important: Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of 200 .

Creating a Request

deleteMessageBatch

Arguments

:: Text
dmbQueueURL

-> DeleteMessageBatch

Creates a value of DeleteMessageBatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmbQueueURL - The URL of the Amazon SQS queue from which messages are deleted. Queue URLs are case-sensitive.

dmbEntries - A list of receipt handles for the messages to be deleted.

data DeleteMessageBatch

See: deleteMessageBatch smart constructor.

Request Lenses

dmbQueueURL :: Lens' DeleteMessageBatch Text

The URL of the Amazon SQS queue from which messages are deleted. Queue URLs are case-sensitive.

dmbEntries :: Lens' DeleteMessageBatch [DeleteMessageBatchRequestEntry]

A list of receipt handles for the messages to be deleted.

Destructuring the Response

deleteMessageBatchResponse

Arguments

:: Int
dmbrsResponseStatus

-> DeleteMessageBatchResponse

Creates a value of DeleteMessageBatchResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmbrsResponseStatus - -- | The response status code.

dmbrsSuccessful - A list of DeleteMessageBatchResultEntry items.

dmbrsFailed - A list of BatchResultErrorEntry items.

data DeleteMessageBatchResponse

For each message in the batch, the response contains a DeleteMessageBatchResultEntry tag if the message is deleted or a BatchResultErrorEntry tag if the message can't be deleted.

See: deleteMessageBatchResponse smart constructor.

Response Lenses

dmbrsResponseStatus :: Lens' DeleteMessageBatchResponse Int

- | The response status code.

dmbrsSuccessful :: Lens' DeleteMessageBatchResponse [DeleteMessageBatchResultEntry]

A list of DeleteMessageBatchResultEntry items.

dmbrsFailed :: Lens' DeleteMessageBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.