Network.AWS.SQS.GetQueueURL

Description

Returns the URL of an existing queue. This action provides a simple way to retrieve the URL of an Amazon SQS queue.

To access a queue that belongs to another AWS account, use the QueueOwnerAWSAccountId parameter to specify the account ID of the queue's owner. The queue's owner must grant you permission to access the queue. For more information about shared queue access, see AddPermission or see Shared Queues in the Amazon SQS Developer Guide .

Creating a Request

getQueueURL

Arguments

:: Text
gquQueueName

-> GetQueueURL

Creates a value of GetQueueURL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gquQueueOwnerAWSAccountId - The AWS account ID of the account that created the queue.

gquQueueName - The name of the queue whose URL must be fetched. Maximum 80 characters. Valid values: alphanumeric characters, hyphens (- ), and underscores (_ ). Queue names are case-sensitive.

data GetQueueURL

See: getQueueURL smart constructor.

Request Lenses

gquQueueOwnerAWSAccountId :: Lens' GetQueueURL (Maybe Text)

The AWS account ID of the account that created the queue.

gquQueueName :: Lens' GetQueueURL Text

The name of the queue whose URL must be fetched. Maximum 80 characters. Valid values: alphanumeric characters, hyphens (- ), and underscores (_ ). Queue names are case-sensitive.

Destructuring the Response

getQueueURLResponse

Arguments

:: Int
gqursResponseStatus

-> Text
gqursQueueURL

-> GetQueueURLResponse

Creates a value of GetQueueURLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gqursResponseStatus - -- | The response status code.

gqursQueueURL - The URL of the queue.

data GetQueueURLResponse

For more information, see Responses in the Amazon SQS Developer Guide .

See: getQueueURLResponse smart constructor.

Response Lenses

gqursResponseStatus :: Lens' GetQueueURLResponse Int

- | The response status code.

gqursQueueURL :: Lens' GetQueueURLResponse Text

The URL of the queue.