Network.AWS.SQS.ListDeadLetterSourceQueues

Description

Returns a list of your queues that have the RedrivePolicy queue attribute configured with a dead letter queue.

For more information about using dead letter queues, see Using Amazon SQS Dead Letter Queues in the Amazon SQS Developer Guide .

Creating a Request

listDeadLetterSourceQueues

Arguments

:: Text
ldlsqQueueURL

-> ListDeadLetterSourceQueues

Creates a value of ListDeadLetterSourceQueues with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldlsqQueueURL - The URL of a dead letter queue. Queue URLs are case-sensitive.

data ListDeadLetterSourceQueues

See: listDeadLetterSourceQueues smart constructor.

Request Lenses

ldlsqQueueURL :: Lens' ListDeadLetterSourceQueues Text

The URL of a dead letter queue. Queue URLs are case-sensitive.

Destructuring the Response

listDeadLetterSourceQueuesResponse

Arguments

:: Int
ldlsqrsResponseStatus

-> ListDeadLetterSourceQueuesResponse

Creates a value of ListDeadLetterSourceQueuesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldlsqrsResponseStatus - -- | The response status code.

ldlsqrsQueueURLs - A list of source queue URLs that have the RedrivePolicy queue attribute configured with a dead letter queue.

data ListDeadLetterSourceQueuesResponse

A list of your dead letter source queues.

See: listDeadLetterSourceQueuesResponse smart constructor.

Response Lenses

ldlsqrsResponseStatus :: Lens' ListDeadLetterSourceQueuesResponse Int

- | The response status code.

ldlsqrsQueueURLs :: Lens' ListDeadLetterSourceQueuesResponse [Text]

A list of source queue URLs that have the RedrivePolicy queue attribute configured with a dead letter queue.