Network.AWS.SQS.ListQueues

Description

Returns a list of your queues. The maximum number of queues that can be returned is 1,000. If you specify a value for the optional QueueNamePrefix parameter, only queues with a name that begins with the specified value are returned.

Creating a Request

listQueues :: ListQueues

Creates a value of ListQueues with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lqQueueNamePrefix - A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned. Queue names are case-sensitive.

data ListQueues

See: listQueues smart constructor.

Request Lenses

lqQueueNamePrefix :: Lens' ListQueues (Maybe Text)

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned. Queue names are case-sensitive.

Destructuring the Response

listQueuesResponse

Arguments

:: Int
lqrsResponseStatus

-> ListQueuesResponse

Creates a value of ListQueuesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lqrsQueueURLs - A list of queue URLs, up to 1,000 entries.

lqrsResponseStatus - -- | The response status code.

data ListQueuesResponse

A list of your queues.

See: listQueuesResponse smart constructor.

Response Lenses

lqrsQueueURLs :: Lens' ListQueuesResponse [Text]

A list of queue URLs, up to 1,000 entries.

lqrsResponseStatus :: Lens' ListQueuesResponse Int

- | The response status code.