Network.AWS.SQS.RemovePermission

Description

Revokes any permissions in the queue policy that matches the specified Label parameter. Only the owner of the queue can remove permissions.

Creating a Request

removePermission

Arguments

:: Text
rpQueueURL

-> Text
rpLabel

-> RemovePermission

Creates a value of RemovePermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpQueueURL - The URL of the Amazon SQS queue from which permissions are removed. Queue URLs are case-sensitive.

rpLabel - The identification of the permission to remove. This is the label added using the AddPermission action.

data RemovePermission

See: removePermission smart constructor.

Request Lenses

rpQueueURL :: Lens' RemovePermission Text

The URL of the Amazon SQS queue from which permissions are removed. Queue URLs are case-sensitive.

rpLabel :: Lens' RemovePermission Text

The identification of the permission to remove. This is the label added using the AddPermission action.

Destructuring the Response

removePermissionResponse :: RemovePermissionResponse

Creates a value of RemovePermissionResponse with the minimum fields required to make a request.

data RemovePermissionResponse

See: removePermissionResponse smart constructor.