Network.AWS.SQS.SendMessage

Description

Delivers a message to the specified queue.

Important: A message can include only XML, JSON, and unformatted text. The following Unicode characters are allowed:

#x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF

Any characters not included in this list will be rejected. For more information, see the W3C specification for characters .

Creating a Request

sendMessage

Arguments

:: Text
smQueueURL

-> Text
smMessageBody

-> SendMessage

Creates a value of SendMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smMessageAttributes - Each message attribute consists of a Name , Type , and Value . For more information, see Message Attribute Items and Validation in the Amazon SQS Developer Guide .

smDelaySeconds - The length of time, in seconds, for which to delay a specific message. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue applies.

smMessageDeduplicationId - This parameter applies only to FIFO (first-in-first-out) queues. The token used for deduplication of sent messages. If a message with a particular MessageDeduplicationId is sent successfully, any messages sent with the same MessageDeduplicationId are accepted successfully but aren't delivered during the 5-minute deduplication interval. For more information, see Exactly-Once Processing in the Amazon SQS Developer Guide . * Every message must have a unique MessageDeduplicationId , * You may provide a MessageDeduplicationId explicitly. * If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message). * If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error. * If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one. * When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered. * If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId , the two messages are treated as duplicates and only one copy of the message is delivered. The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z , A-Z , 0-9 ) and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~ ). For best practices of using MessageDeduplicationId@ , see Using the MessageDeduplicationId Property in the Amazon Simple Queue Service Developer Guide .

smMessageGroupId - This parameter applies only to FIFO (first-in-first-out) queues. The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple readers can process the queue, but the session data of each user is processed in a FIFO fashion. * You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId , the action fails. * ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId , the messages are sorted by time sent. The caller can't specify a MessageGroupId . The length of MessageGroupId is 128 characters. Valid values are alphanumeric characters and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~) . For best practices of using MessageGroupId , see Using the MessageGroupId Property in the Amazon Simple Queue Service Developer Guide . Important: MessageGroupId@ is required for FIFO queues. You can't use it for Standard queues.

smQueueURL - The URL of the Amazon SQS queue to which a message is sent. Queue URLs are case-sensitive.

smMessageBody - The message to send. The maximum string size is 256 KB. Important: A message can include only XML, JSON, and unformatted text. The following Unicode characters are allowed: #x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF Any characters not included in this list will be rejected. For more information, see the W3C specification for characters .

data SendMessage

See: sendMessage smart constructor.

Request Lenses

smMessageAttributes :: Lens' SendMessage (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name , Type , and Value . For more information, see Message Attribute Items and Validation in the Amazon SQS Developer Guide .

smDelaySeconds :: Lens' SendMessage (Maybe Int)

The length of time, in seconds, for which to delay a specific message. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue applies.

smMessageDeduplicationId :: Lens' SendMessage (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues. The token used for deduplication of sent messages. If a message with a particular MessageDeduplicationId is sent successfully, any messages sent with the same MessageDeduplicationId are accepted successfully but aren't delivered during the 5-minute deduplication interval. For more information, see Exactly-Once Processing in the Amazon SQS Developer Guide . * Every message must have a unique MessageDeduplicationId , * You may provide a MessageDeduplicationId explicitly. * If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message). * If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error. * If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one. * When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered. * If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId , the two messages are treated as duplicates and only one copy of the message is delivered. The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z , A-Z , 0-9 ) and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~ ). For best practices of using MessageDeduplicationId@ , see Using the MessageDeduplicationId Property in the Amazon Simple Queue Service Developer Guide .

smMessageGroupId :: Lens' SendMessage (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues. The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple readers can process the queue, but the session data of each user is processed in a FIFO fashion. * You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId , the action fails. * ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId , the messages are sorted by time sent. The caller can't specify a MessageGroupId . The length of MessageGroupId is 128 characters. Valid values are alphanumeric characters and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~) . For best practices of using MessageGroupId , see Using the MessageGroupId Property in the Amazon Simple Queue Service Developer Guide . Important: MessageGroupId@ is required for FIFO queues. You can't use it for Standard queues.

smQueueURL :: Lens' SendMessage Text

The URL of the Amazon SQS queue to which a message is sent. Queue URLs are case-sensitive.

smMessageBody :: Lens' SendMessage Text

The message to send. The maximum string size is 256 KB. Important: A message can include only XML, JSON, and unformatted text. The following Unicode characters are allowed: #x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF Any characters not included in this list will be rejected. For more information, see the W3C specification for characters .

Destructuring the Response

sendMessageResponse

Arguments

:: Int
smrsResponseStatus

-> SendMessageResponse

Creates a value of SendMessageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smrsSequenceNumber - This parameter applies only to FIFO (first-in-first-out) queues. The large, non-consecutive number that Amazon SQS assigns to each message. The length of SequenceNumber is 128 bits. SequenceNumber continues to increase for a particular MessageGroupId .

smrsMessageId - An attribute containing the MessageId of the message sent to the queue. For more information, see Queue and Message Identifiers in the Amazon SQS Developer Guide .

smrsMD5OfMessageBody - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

smrsMD5OfMessageAttributes - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

smrsResponseStatus - -- | The response status code.

data SendMessageResponse

The MD5OfMessageBody and MessageId elements.

See: sendMessageResponse smart constructor.

Response Lenses

smrsSequenceNumber :: Lens' SendMessageResponse (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues. The large, non-consecutive number that Amazon SQS assigns to each message. The length of SequenceNumber is 128 bits. SequenceNumber continues to increase for a particular MessageGroupId .

smrsMessageId :: Lens' SendMessageResponse (Maybe Text)

An attribute containing the MessageId of the message sent to the queue. For more information, see Queue and Message Identifiers in the Amazon SQS Developer Guide .

smrsMD5OfMessageBody :: Lens' SendMessageResponse (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

smrsMD5OfMessageAttributes :: Lens' SendMessageResponse (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

smrsResponseStatus :: Lens' SendMessageResponse Int

- | The response status code.