Network.AWS.SQS.SendMessageBatch

Description

Delivers up to ten messages to the specified queue. This is a batch version of SendMessage . For a FIFO queue, multiple messages within a single batch are enqueued in the order they are sent.

The result of sending each message is reported individually in the response. Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of 200 .

The maximum allowed individual message size and the maximum total payload size (the sum of the individual lengths of all of the batched messages) are both 256 KB (262,144 bytes).

Important: A message can include only XML, JSON, and unformatted text. The following Unicode characters are allowed:

#x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF

Any characters not included in this list will be rejected. For more information, see the W3C specification for characters .

If you don't specify the DelaySeconds parameter for an entry, Amazon SQS uses the default value for the queue.

Creating a Request

sendMessageBatch

Arguments

:: Text
smbQueueURL

-> SendMessageBatch

Creates a value of SendMessageBatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smbQueueURL - The URL of the Amazon SQS queue to which batched messages are sent. Queue URLs are case-sensitive.

smbEntries - A list of SendMessageBatchRequestEntry items.

data SendMessageBatch

See: sendMessageBatch smart constructor.

Request Lenses

smbQueueURL :: Lens' SendMessageBatch Text

The URL of the Amazon SQS queue to which batched messages are sent. Queue URLs are case-sensitive.

smbEntries :: Lens' SendMessageBatch [SendMessageBatchRequestEntry]

A list of SendMessageBatchRequestEntry items.

Destructuring the Response

sendMessageBatchResponse

Arguments

:: Int
smbrsResponseStatus

-> SendMessageBatchResponse

Creates a value of SendMessageBatchResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smbrsResponseStatus - -- | The response status code.

smbrsSuccessful - A list of SendMessageBatchResultEntry items.

smbrsFailed - A list of BatchResultErrorEntry items with error details about each message that can't be enqueued.

data SendMessageBatchResponse

For each message in the batch, the response contains a SendMessageBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

See: sendMessageBatchResponse smart constructor.

Response Lenses

smbrsResponseStatus :: Lens' SendMessageBatchResponse Int

- | The response status code.

smbrsSuccessful :: Lens' SendMessageBatchResponse [SendMessageBatchResultEntry]

A list of SendMessageBatchResultEntry items.

smbrsFailed :: Lens' SendMessageBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items with error details about each message that can't be enqueued.