Network.AWS.SQS.Types

Description

Service Configuration

sqs :: Service

API version 2012-11-05 of the Amazon Simple Queue Service SDK configuration.

Errors

_InvalidBatchEntryId :: AsError a => Getting (First ServiceError) a ServiceError

The Id of a batch entry in a batch request doesn't abide by the specification.

_TooManyEntriesInBatchRequest :: AsError a => Getting (First ServiceError) a ServiceError

The batch request contains more entries than permissible.

_QueueDeletedRecently :: AsError a => Getting (First ServiceError) a ServiceError

You must wait 60 seconds after deleting a queue before you can create another one with the same name.

_QueueDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The queue referred to doesn't exist.

_InvalidAttributeName :: AsError a => Getting (First ServiceError) a ServiceError

The attribute referred to doesn't exist.

_UnsupportedOperation :: AsError a => Getting (First ServiceError) a ServiceError

Error code 400. Unsupported operation.

_InvalidMessageContents :: AsError a => Getting (First ServiceError) a ServiceError

The message contains characters outside the allowed set.

_BatchRequestTooLong :: AsError a => Getting (First ServiceError) a ServiceError

The length of all the messages put together is more than the limit.

_OverLimit :: AsError a => Getting (First ServiceError) a ServiceError

The action that you requested would violate a limit. For example, ReceiveMessage returns this error if the maximum number of inflight messages is reached. AddPermission returns this error if the maximum number of permissions for the queue is reached.

_QueueNameExists :: AsError a => Getting (First ServiceError) a ServiceError

A queue already exists with this name. Amazon SQS returns this error only if the request includes attributes whose values differ from those of the existing queue.

_PurgeQueueInProgress :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the specified queue previously received a PurgeQueue request within the last 60 seconds (the time it can take to delete the messages in the queue).

_InvalidIdFormat :: AsError a => Getting (First ServiceError) a ServiceError

The receipt handle isn't valid for the current version.

_ReceiptHandleIsInvalid :: AsError a => Getting (First ServiceError) a ServiceError

The receipt handle provided isn't valid.

_EmptyBatchRequest :: AsError a => Getting (First ServiceError) a ServiceError

The batch request doesn't contain any entries.

_BatchEntryIdsNotDistinct :: AsError a => Getting (First ServiceError) a ServiceError

Two or more batch entries in the request have the same Id .

_MessageNotInflight :: AsError a => Getting (First ServiceError) a ServiceError

The message referred to isn't in flight.

MessageAttribute

data MessageAttribute

Constructors

All

ApproximateFirstReceiveTimestamp

ApproximateReceiveCount

SenderId

SentTimestamp

QueueAttributeName

data QueueAttributeName

Constructors

QANAll

QANApproximateNumberOfMessages

QANApproximateNumberOfMessagesDelayed

QANApproximateNumberOfMessagesNotVisible

QANContentBasedDeduplication

QANCreatedTimestamp

QANDelaySeconds

QANFifoQueue

QANKMSDataKeyReusePeriodSeconds

QANKMSMasterKeyId

QANLastModifiedTimestamp

QANMaximumMessageSize

QANMessageRetentionPeriod

QANPolicy

QANQueueARN

QANReceiveMessageWaitTimeSeconds

QANRedrivePolicy

QANVisibilityTimeout

BatchResultErrorEntry

data BatchResultErrorEntry

This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

See: batchResultErrorEntry smart constructor.

batchResultErrorEntry

Arguments

:: Text
breeId

-> Bool
breeSenderFault

-> Text
breeCode

-> BatchResultErrorEntry

Creates a value of BatchResultErrorEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

breeMessage - A message explaining why the action failed on this entry.

breeId - The Id of an entry in a batch request.

breeSenderFault - Specifies whether the error happened due to the sender's fault.

breeCode - An error code representing why the action failed on this entry.

breeMessage :: Lens' BatchResultErrorEntry (Maybe Text)

A message explaining why the action failed on this entry.

breeId :: Lens' BatchResultErrorEntry Text

The Id of an entry in a batch request.

breeSenderFault :: Lens' BatchResultErrorEntry Bool

Specifies whether the error happened due to the sender's fault.

breeCode :: Lens' BatchResultErrorEntry Text

An error code representing why the action failed on this entry.

ChangeMessageVisibilityBatchRequestEntry

data ChangeMessageVisibilityBatchRequestEntry

Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch .

Important: All of the following list parameters must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n , where n is an integer value starting with 1 . For example, a parameter list for this action might look like this:

&amp;ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2

&amp;ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=replaceableYour_Receipt_Handle/replaceable

&amp;ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45

See: changeMessageVisibilityBatchRequestEntry smart constructor.

changeMessageVisibilityBatchRequestEntry

Arguments

:: Text
cId

-> Text
cReceiptHandle

-> ChangeMessageVisibilityBatchRequestEntry

Creates a value of ChangeMessageVisibilityBatchRequestEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cVisibilityTimeout - The new value (in seconds) for the message's visibility timeout.

cId - An identifier for this particular receipt handle used to communicate the result.

cReceiptHandle - A receipt handle.

cVisibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Int)

The new value (in seconds) for the message's visibility timeout.

cId :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

An identifier for this particular receipt handle used to communicate the result.

cReceiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

A receipt handle.

ChangeMessageVisibilityBatchResultEntry

data ChangeMessageVisibilityBatchResultEntry

Encloses the Id of an entry in ChangeMessageVisibilityBatch .

See: changeMessageVisibilityBatchResultEntry smart constructor.

changeMessageVisibilityBatchResultEntry

Arguments

:: Text
cmvbreId

-> ChangeMessageVisibilityBatchResultEntry

Creates a value of ChangeMessageVisibilityBatchResultEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmvbreId - Represents a message whose visibility timeout has been changed successfully.

cmvbreId :: Lens' ChangeMessageVisibilityBatchResultEntry Text

Represents a message whose visibility timeout has been changed successfully.

DeleteMessageBatchRequestEntry

data DeleteMessageBatchRequestEntry

Encloses a receipt handle and an identifier for it.

See: deleteMessageBatchRequestEntry smart constructor.

deleteMessageBatchRequestEntry

Arguments

:: Text
dmbreId

-> Text
dmbreReceiptHandle

-> DeleteMessageBatchRequestEntry

Creates a value of DeleteMessageBatchRequestEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmbreId - An identifier for this particular receipt handle. This is used to communicate the result.

dmbreReceiptHandle - A receipt handle.

dmbreId :: Lens' DeleteMessageBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result.

dmbreReceiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

A receipt handle.

DeleteMessageBatchResultEntry

data DeleteMessageBatchResultEntry

Encloses the Id of an entry in DeleteMessageBatch .

See: deleteMessageBatchResultEntry smart constructor.

deleteMessageBatchResultEntry

Arguments

:: Text
dId

-> DeleteMessageBatchResultEntry

Creates a value of DeleteMessageBatchResultEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dId - Represents a successfully deleted message.

dId :: Lens' DeleteMessageBatchResultEntry Text

Represents a successfully deleted message.

Message

data Message

An Amazon SQS message.

See: message smart constructor.

message :: Message

Creates a value of Message with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mMessageAttributes - Each message attribute consists of a Name , Type , and Value . For more information, see Message Attribute Items and Validation in the Amazon SQS Developer Guide .

mMD5OfBody - An MD5 digest of the non-URL-encoded message body string.

mBody - The message's contents (not URL-encoded).

mAttributes - SenderId , SentTimestamp , ApproximateReceiveCount , andor ApproximateFirstReceiveTimestamp . SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the <http:en.wikipedia.orgwiki/Unix_time epoch time> in milliseconds.

mReceiptHandle - An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

mMessageId - A unique identifier for the message. A MessageId is considered unique across all AWS accounts for an extended period of time.

mMD5OfMessageAttributes - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

mMessageAttributes :: Lens' Message (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name , Type , and Value . For more information, see Message Attribute Items and Validation in the Amazon SQS Developer Guide .

mMD5OfBody :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message body string.

mBody :: Lens' Message (Maybe Text)

The message's contents (not URL-encoded).

mAttributes :: Lens' Message (HashMap MessageAttribute Text)

SenderId , SentTimestamp , ApproximateReceiveCount , andor ApproximateFirstReceiveTimestamp . SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the <http:en.wikipedia.orgwiki/Unix_time epoch time> in milliseconds.

mReceiptHandle :: Lens' Message (Maybe Text)

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

mMessageId :: Lens' Message (Maybe Text)

A unique identifier for the message. A MessageId is considered unique across all AWS accounts for an extended period of time.

mMD5OfMessageAttributes :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the Value attribute has the same restrictions on the content as the message body. For more information, see SendMessage .

Name , type , value and the message body must not be empty or null. All parts of the message attribute, including Name , Type , and Value , are part of the message size restriction (256 KB or 262,144 bytes).

See: messageAttributeValue smart constructor.

messageAttributeValue

Arguments

:: Text
mavDataType

-> MessageAttributeValue

Creates a value of MessageAttributeValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mavBinaryValue - Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

mavStringListValues - Not implemented. Reserved for future use.

mavStringValue - Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters .

mavBinaryListValues - Not implemented. Reserved for future use.

mavDataType - Amazon SQS supports the following logical data types: String , Number , and Binary . For the Number data type, you must use StringValue . You can also append custom labels. For more information, see Message Attribute Data Types and Validation in the Amazon SQS Developer Guide .

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

mavStringListValues :: Lens' MessageAttributeValue [Text]

Not implemented. Reserved for future use.

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters .

mavBinaryListValues :: Lens' MessageAttributeValue [ByteString]

Not implemented. Reserved for future use.

mavDataType :: Lens' MessageAttributeValue Text

Amazon SQS supports the following logical data types: String , Number , and Binary . For the Number data type, you must use StringValue . You can also append custom labels. For more information, see Message Attribute Data Types and Validation in the Amazon SQS Developer Guide .

SendMessageBatchRequestEntry

data SendMessageBatchRequestEntry

Contains the details of a single Amazon SQS message along with an Id .

See: sendMessageBatchRequestEntry smart constructor.

sendMessageBatchRequestEntry

Arguments

:: Text
sId

-> Text
sMessageBody

-> SendMessageBatchRequestEntry

Creates a value of SendMessageBatchRequestEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sMessageAttributes - Each message attribute consists of a Name , Type , and Value . For more information, see Message Attribute Items and Validation in the Amazon SQS Developer Guide .

sDelaySeconds - The length of time, in seconds, for which a specific message is delayed. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue is applied.

sMessageDeduplicationId - This parameter applies only to FIFO (first-in-first-out) queues. The token used for deduplication of messages within a 5-minute minimum deduplication interval. If a message with a particular MessageDeduplicationId is sent successfully, subsequent messages with the same MessageDeduplicationId are accepted successfully but aren't delivered. For more information, see Exactly-Once Processing in the Amazon SQS Developer Guide . * Every message must have a unique MessageDeduplicationId , * You may provide a MessageDeduplicationId explicitly. * If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message). * If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error. * If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one. * When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered. * If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId , the two messages are treated as duplicates and only one copy of the message is delivered. The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z , A-Z , 0-9 ) and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~ ). For best practices of using MessageDeduplicationId@ , see Using the MessageDeduplicationId Property in the Amazon Simple Queue Service Developer Guide .

sMessageGroupId - This parameter applies only to FIFO (first-in-first-out) queues. The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple readers can process the queue, but the session data of each user is processed in a FIFO fashion. * You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId , the action fails. * ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId , the messages are sorted by time sent. The caller can't specify a MessageGroupId . The length of MessageGroupId is 128 characters. Valid values are alphanumeric characters and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~) . For best practices of using MessageGroupId , see Using the MessageGroupId Property in the Amazon Simple Queue Service Developer Guide . Important: MessageGroupId@ is required for FIFO queues. You can't use it for Standard queues.

sId - An identifier for a message in this batch used to communicate the result.

sMessageBody - The body of the message.

sMessageAttributes :: Lens' SendMessageBatchRequestEntry (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name , Type , and Value . For more information, see Message Attribute Items and Validation in the Amazon SQS Developer Guide .

sDelaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Int)

The length of time, in seconds, for which a specific message is delayed. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue is applied.

sMessageDeduplicationId :: Lens' SendMessageBatchRequestEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues. The token used for deduplication of messages within a 5-minute minimum deduplication interval. If a message with a particular MessageDeduplicationId is sent successfully, subsequent messages with the same MessageDeduplicationId are accepted successfully but aren't delivered. For more information, see Exactly-Once Processing in the Amazon SQS Developer Guide . * Every message must have a unique MessageDeduplicationId , * You may provide a MessageDeduplicationId explicitly. * If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message). * If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error. * If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one. * When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered. * If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId , the two messages are treated as duplicates and only one copy of the message is delivered. The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z , A-Z , 0-9 ) and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~ ). For best practices of using MessageDeduplicationId@ , see Using the MessageDeduplicationId Property in the Amazon Simple Queue Service Developer Guide .

sMessageGroupId :: Lens' SendMessageBatchRequestEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues. The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple readers can process the queue, but the session data of each user is processed in a FIFO fashion. * You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId , the action fails. * ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId , the messages are sorted by time sent. The caller can't specify a MessageGroupId . The length of MessageGroupId is 128 characters. Valid values are alphanumeric characters and punctuation (!"#$%&'()*+,-./:;=?[]^_`{|}~) . For best practices of using MessageGroupId , see Using the MessageGroupId Property in the Amazon Simple Queue Service Developer Guide . Important: MessageGroupId@ is required for FIFO queues. You can't use it for Standard queues.

sId :: Lens' SendMessageBatchRequestEntry Text

An identifier for a message in this batch used to communicate the result.

sMessageBody :: Lens' SendMessageBatchRequestEntry Text

The body of the message.

SendMessageBatchResultEntry

data SendMessageBatchResultEntry

Encloses a MessageId for a successfully-enqueued message in a SendMessageBatch .

See: sendMessageBatchResultEntry smart constructor.

sendMessageBatchResultEntry

Arguments

:: Text
smbreId

-> Text
smbreMessageId

-> Text
smbreMD5OfMessageBody

-> SendMessageBatchResultEntry

Creates a value of SendMessageBatchResultEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smbreSequenceNumber - This parameter applies only to FIFO (first-in-first-out) queues. The large, non-consecutive number that Amazon SQS assigns to each message. The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId .

smbreMD5OfMessageAttributes - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

smbreId - An identifier for the message in this batch.

smbreMessageId - An identifier for the message.

smbreMD5OfMessageBody - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

smbreSequenceNumber :: Lens' SendMessageBatchResultEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues. The large, non-consecutive number that Amazon SQS assigns to each message. The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId .

smbreMD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .

smbreId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message in this batch.

smbreMessageId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message.

smbreMD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321 .