Network.AWS.SQS.Types

Service

data SQS

Version 2012-11-05 of the Amazon Simple Queue Service service.

Error

data RESTError :: *

XML

ns :: Text

DeleteMessageBatchRequestEntry

data DeleteMessageBatchRequestEntry

deleteMessageBatchRequestEntry

Arguments

:: Text
dmbreId

-> Text
dmbreReceiptHandle

-> DeleteMessageBatchRequestEntry

DeleteMessageBatchRequestEntry constructor.

The fields accessible through corresponding lenses are:

dmbreId :: Text

dmbreReceiptHandle :: Text

dmbreId :: Lens' DeleteMessageBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

dmbreReceiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

A receipt handle.

MessageAttributeValue

data MessageAttributeValue

messageAttributeValue

Arguments

:: Text
mavDataType

-> MessageAttributeValue

MessageAttributeValue constructor.

The fields accessible through corresponding lenses are:

mavBinaryListValues :: [Base64]

mavBinaryValue :: Maybe Base64

mavDataType :: Text

mavStringListValues :: [Text]

mavStringValue :: Maybe Text

mavBinaryListValues :: Lens' MessageAttributeValue [Base64]

Not implemented. Reserved for future use.

mavBinaryValue :: Lens' MessageAttributeValue (Maybe Base64)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

mavDataType :: Lens' MessageAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. In addition, you can append your own custom labels. For more information, see Message Attribute Data Types.

mavStringListValues :: Lens' MessageAttributeValue [Text]

Not implemented. Reserved for future use.

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http://en.wikipedia.org/wiki/ASCII#ASCII_printable_characters.

ChangeMessageVisibilityBatchResultEntry

data ChangeMessageVisibilityBatchResultEntry

changeMessageVisibilityBatchResultEntry

Arguments

:: Text
cmvbreId

-> ChangeMessageVisibilityBatchResultEntry

ChangeMessageVisibilityBatchResultEntry constructor.

The fields accessible through corresponding lenses are:

cmvbreId :: Text

cmvbreId :: Lens' ChangeMessageVisibilityBatchResultEntry Text

Represents a message whose visibility timeout has been changed successfully.

ChangeMessageVisibilityBatchRequestEntry

data ChangeMessageVisibilityBatchRequestEntry

changeMessageVisibilityBatchRequestEntry

Arguments

:: Text
cmvbre1Id

-> Text
cmvbre1ReceiptHandle

-> ChangeMessageVisibilityBatchRequestEntry

ChangeMessageVisibilityBatchRequestEntry constructor.

The fields accessible through corresponding lenses are:

cmvbre1Id :: Text

cmvbre1ReceiptHandle :: Text

cmvbre1VisibilityTimeout :: Maybe Int

cmvbre1Id :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

cmvbre1ReceiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

A receipt handle.

cmvbre1VisibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Int)

The new value (in seconds) for the message's visibility timeout.

DeleteMessageBatchResultEntry

data DeleteMessageBatchResultEntry

deleteMessageBatchResultEntry

Arguments

:: Text
dmbre1Id

-> DeleteMessageBatchResultEntry

DeleteMessageBatchResultEntry constructor.

The fields accessible through corresponding lenses are:

dmbre1Id :: Text

dmbre1Id :: Lens' DeleteMessageBatchResultEntry Text

Represents a successfully deleted message.

Message

data Message

message :: Message

Message constructor.

The fields accessible through corresponding lenses are:

mAttributes :: HashMap Text Text

mBody :: Maybe Text

mMD5OfBody :: Maybe Text

mMD5OfMessageAttributes :: Maybe Text

mMessageAttributes :: HashMap Text MessageAttributeValue

mMessageId :: Maybe Text

mReceiptHandle :: Maybe Text

mAttributes :: Lens' Message (HashMap Text Text)

SenderId, SentTimestamp, ApproximateReceiveCount, and/or ApproximateFirstReceiveTimestamp. SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the epoch time in milliseconds.

mBody :: Lens' Message (Maybe Text)

The message's contents (not URL-encoded).

mMD5OfBody :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message body string.

mMD5OfMessageAttributes :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http://www.faqs.org/rfcs/rfc1321.html.

mMessageAttributes :: Lens' Message (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

mMessageId :: Lens' Message (Maybe Text)

A unique identifier for the message. Message IDs are considered unique across all AWS accounts for an extended period of time.

mReceiptHandle :: Lens' Message (Maybe Text)

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

SendMessageBatchRequestEntry

data SendMessageBatchRequestEntry

sendMessageBatchRequestEntry

Arguments

:: Text
smbreId

-> Text
smbreMessageBody

-> SendMessageBatchRequestEntry

SendMessageBatchRequestEntry constructor.

The fields accessible through corresponding lenses are:

smbreDelaySeconds :: Maybe Int

smbreId :: Text

smbreMessageAttributes :: HashMap Text MessageAttributeValue

smbreMessageBody :: Text

smbreDelaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Int)

The number of seconds for which the message has to be delayed.

smbreId :: Lens' SendMessageBatchRequestEntry Text

An identifier for the message in this batch. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

smbreMessageAttributes :: Lens' SendMessageBatchRequestEntry (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

smbreMessageBody :: Lens' SendMessageBatchRequestEntry Text

Body of the message.

SendMessageBatchResultEntry

data SendMessageBatchResultEntry

sendMessageBatchResultEntry

Arguments

:: Text
smbre1Id

-> Text
smbre1MessageId

-> Text
smbre1MD5OfMessageBody

-> SendMessageBatchResultEntry

SendMessageBatchResultEntry constructor.

The fields accessible through corresponding lenses are:

smbre1Id :: Text

smbre1MD5OfMessageAttributes :: Maybe Text

smbre1MD5OfMessageBody :: Text

smbre1MessageId :: Text

smbre1Id :: Lens' SendMessageBatchResultEntry Text

An identifier for the message in this batch.

smbre1MD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message batch correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http://www.faqs.org/rfcs/rfc1321.html.

smbre1MD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http://www.faqs.org/rfcs/rfc1321.html.

smbre1MessageId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message.

BatchResultErrorEntry

data BatchResultErrorEntry

batchResultErrorEntry

Arguments

:: Text
breeId

-> Bool
breeSenderFault

-> Text
breeCode

-> BatchResultErrorEntry

BatchResultErrorEntry constructor.

The fields accessible through corresponding lenses are:

breeCode :: Text

breeId :: Text

breeMessage :: Maybe Text

breeSenderFault :: Bool

breeCode :: Lens' BatchResultErrorEntry Text

An error code representing why the action failed on this entry.

breeId :: Lens' BatchResultErrorEntry Text

The id of an entry in a batch request.

breeMessage :: Lens' BatchResultErrorEntry (Maybe Text)

A message explaining why the action failed on this entry.

breeSenderFault :: Lens' BatchResultErrorEntry Bool

Whether the error happened due to the sender's fault.