Network.AWS.SQS.V2012_11_05.AddPermission

Description

Adds a permission to a queue for a specific principal. This allows for sharing access to the queue. When you create a queue, you have full control access rights for the queue. Only you (as owner of the queue) can grant or deny permissions to the queue. For more information about these permissions, see Shared Queues in the Amazon SQS Developer Guide. AddPermission writes an Amazon SQS-generated policy. If you want to write your own policy, use SetQueueAttributes to upload your policy. For more information about writing your own policy, see Using The Access Policy Language in the Amazon SQS Developer Guide. Some API actions take lists of parameters. These lists are specified using the param.n notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this: &amp;Attribute.1=this &amp;Attribute.2=that The following example Query request grants a SendMessage permission to the principal whose AWS account number is 125074342641. http:/sqs.us-east-1.amazonaws.com123456789012testQueue ?Action=AddPermission &Label=testLabel &AWSAccountId.1=125074342641 &ActionName.1=SendMessage &AWSAccountId.2=125074342642 &ActionName.2=ReceiveMessage &Version=2009-02-01 &SignatureMethod=HmacSHA256 &Expires=2009-04-18T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE 9a285199-c8d6-47c2-bdb2-314cb47d599d.

Request

data AddPermission

Request constructor

mkAddPermission

Arguments

:: Text
apQueueUrl

-> Text
apLabel

-> [Text]
apAWSAccountIds

-> [Text]
apActions

-> AddPermission

Smart constructor for the minimum required parameters to construct a valid AddPermission request.

Request lenses

apQueueUrl :: Lens' AddPermission Text

The URL of the Amazon SQS queue to take action on.

apLabel :: Lens' AddPermission Text

The unique identification of the permission you're setting (e.g., AliceSendMessage). Constraints: Maximum 80 characters; alphanumeric characters, hyphens (-), and underscores (_) are allowed.

apAWSAccountIds :: Lens' AddPermission [Text]

The AWS account number of the principal who will be given permission. The principal must have an AWS account, but does not need to be signed up for Amazon SQS. For information about locating the AWS account identification, see Your AWS Identifiers in the Amazon SQS Developer Guide.

apActions :: Lens' AddPermission [Text]

The action the client wants to allow for the specified principal. The following are valid values: * | SendMessage | ReceiveMessage | DeleteMessage | ChangeMessageVisibility | GetQueueAttributes | GetQueueUrl. For more information about these actions, see Understanding Permissions in the Amazon SQS Developer Guide. Specifying SendMessage, DeleteMessage, or ChangeMessageVisibility for the ActionName.n also grants permissions for the corresponding batch versions of those actions: SendMessageBatch, DeleteMessageBatch, and ChangeMessageVisibilityBatch.

Response

data AddPermissionResponse