Network.AWS.SQS.V2012_11_05.ListQueues

Description

Returns a list of your queues. The maximum number of queues that can be returned is 1000. If you specify a value for the optional QueueNamePrefix parameter, only queues with a name beginning with the specified value are returned. The following example Query request returns the queues whose names begin with the letter T. http:/sqs.us-east-1.amazonaws.com ?Action=ListQueues &QueueNamePrefix=t &Version=2009-02-01 &SignatureMethod=HmacSHA256 &Expires=2009-04-18T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE http:/sqs.us-east-1.amazonaws.com123456789012/testQueue 725275ae-0b9b-4762-b238-436d7c65a1ac.

Documentation

listQueues :: ListQueues

Minimum specification for a ListQueues request.

data ListQueues

Constructors

ListQueues

Fields

_lqrQueueNamePrefix :: Maybe Text

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

lqrQueueNamePrefix :: Iso' ListQueues (Maybe Text)

data ListQueuesResponse

Constructors

ListQueuesResponse

Fields

_lqsQueueUrls :: [Text]

A list of queue URLs, up to 1000 entries.

lqsQueueUrls :: Iso' ListQueuesResponse [Text]