Network.AWS.SQS.V2012_11_05.SendMessageBatch

Description

Delivers up to ten messages to the specified queue. This is a batch version of SendMessage. The result of the send action on each message is reported individually in the response. The maximum allowed individual message size is 256 KB (262,144 bytes). The maximum total payload size (i.e., the sum of all a batch's individual message lengths) is also 256 KB (262,144 bytes). If the DelaySeconds parameter is not specified for an entry, the default for the queue is used. The following list shows the characters (in Unicode) that are allowed in your message, according to the W3C XML specification. For more information, go to http:/www.faqs.orgrfcs/rfc1321.html. If you send any characters that are not included in the list, your request will be rejected. #x9 | #xA | #xD | [#x20 to #xD7FF] | [#xE000 to #xFFFD] | [#x10000 to #x10FFFF] Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of 200. Some API actions take lists of parameters. These lists are specified using the param.n notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this: &amp;Attribute.1=this &amp;Attribute.2=that The following example SendMessageBatch request sends two messages to the queue. You must URL encode the entire URL; however, we've URL encoded only the message body to make the example easier for you to read. The following example response includes the MD5 digest for the messages. http:/sqs.us-east-1.amazonaws.com123456789012testQueue ?Action=SendMessageBatch &SendMessageBatchRequestEntry.1.Id=test_msg_001 &SendMessageBatchRequestEntry.1.MessageBody=test%20message%20body%201 &SendMessageBatchRequestEntry.2.Id=test_msg_002 &SendMessageBatchRequestEntry.2.MessageBody=test%20message%20body%202 &SendMessageBatchRequestEntry.2.DelaySeconds=60 &SendMessageBatchRequestEntry.2.MessageAttribute.1.Name=test_attribute_name_1

&SendMessageBatchRequestEntry.2.MessageAttribute.1.Value.StringValue=test_attribute_value_1 &SendMessageBatchRequestEntry.2.MessageAttribute.1.Value.DataType=String &Version=2012-11-05 &SignatureMethod=HmacSHA256 &Expires=2014-05-05T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE test_msg_001 0a5231c7-8bff-4955-be2e-8dc7c50a25fa 0e024d309850c78cba5eabbeff7cae71 test_msg_002 15ee1ed3-87e7-40c1-bdaa-2e49968ea7e9 7fb8146a82f95e0af155278f406862c2 295c5fa15a51aae6884d1d7c1d99ca50 ca1ad5d0-8271-408b-8d0f-1351bf547e74.

Request

data SendMessageBatch

Request constructor

mkSendMessageBatch

Arguments

:: Text
smbQueueUrl

-> [SendMessageBatchRequestEntry]
smbEntries

-> SendMessageBatch

Smart constructor for the minimum required parameters to construct a valid SendMessageBatch request.

Request lenses

smbQueueUrl :: Lens' SendMessageBatch Text

The URL of the Amazon SQS queue to take action on.

smbEntries :: Lens' SendMessageBatch [SendMessageBatchRequestEntry]

A list of SendMessageBatchRequestEntry items.

Response

data SendMessageBatchResponse

For each message in the batch, the response contains a SendMessageBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

Response lenses

smbrSuccessful :: Lens' SendMessageBatchResponse [SendMessageBatchResultEntry]

A list of SendMessageBatchResultEntry items.

smbrFailed :: Lens' SendMessageBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items with the error detail about each message that could not be enqueued.